+++ tests/sketch_tests.cpp
@@ -1,270 +1,26 @@
-#include <gtest/gtest.h>
+#include "sketch_test_fixture.h"
 
 #include <TopoDS.hxx>
 #include <TopoDS_Wire.hxx>
-#include <BRepPrimAPI_MakeRevol.hxx>
-#include <BRepTools.hxx>
-#include <BRepTools_WireExplorer.hxx>
-#include <GProp_GProps.hxx>
-#include <BRepGProp.hxx>
-#include <Bnd_Box.hxx>
-#include <BRepBndLib.hxx>
 #include <algorithm>
+#include <cmath>
 #include <iostream>
 #include <numbers>
 #include <set>
-
+#include <string>
+#include <vector>
 #include "utl_geom.h"
-#include "gui.h"
-#include "gui_occt_view.h"
-#include "sketch.h"
-#include "sketch_ais.h"
-#include "sketch_op_recorder.h"
-#include "sketch_json.h"
 #include "sketch_edge.h"
+#include "sketch_json.h"
 #include "sketch_nodes.h"
-#include "utl_io.h"
-#include "utl_asset_store.h"
+#include "sketch_op_recorder.h"
 #include "utl_occt.h"
+#include <BRepBndLib.hxx>
+#include <Bnd_Box.hxx>
+#include <GProp_GProps.hxx>
+#include <BRepGProp.hxx>
 
 using namespace glm;
-
-// to_wkt_string is provided by geom.h (pure C++ implementation)
-
-class GUI_access
-{
- public:
-  static void        set_view(GUI& gui, std::unique_ptr<Occt_view>& view);
-  static Occt_view&  get_view(GUI& gui);
-  static std::string get_message(const GUI& gui);
-  static void        sketch_left_click(GUI& gui, const ScreenCoords& screen_coords);
-  static void        mirror_selected_edges(GUI& gui);
-};
-
-// Test fixture for Sketch tests
-class Sketch_test : public ::testing::Test
-{
- protected:
-  void SetUp() override
-  {
-    // TODO it is creating a view each time, create only one view and have a reset method.
-    std::unique_ptr<Occt_view> view = std::make_unique<Occt_view>(s_gui);
-    view->init_viewer();
-    view->init_default();
-    GUI_access::set_view(s_gui, view);
-
-    // Force midpoint creation on for tests (they were written assuming auto mids on linear edges).
-    // Production default is off (see user's request); specific tests can override with set_... (false).
-    Sketch::set_add_mid_pt_edges(true);
-  }
-
-  void TearDown() override
-  {
-    // view.reset();
-  }
-
-  Occt_view& view()
-  {
-    return GUI_access::get_view(s_gui);
-  }
-
-  GUI& gui()
-  {
-    return s_gui;
-  }
-
-  // Helper to add all permutations of edges (with both orientations) to a sketch and call a lambda
-  template <typename Callback>
-  void for_all_edge_permutations_(
-      const std::vector<gp_Pnt2d>&            points,
-      const std::vector<std::pair<int, int>>& edge_indices,
-      const gp_Pln&                           plane,
-      Callback&&                              callback);
-
-  // Helper to add edges to a sketch from a vector of points and edge index pairs
-  template <typename Callback>
-  void add_edges_from_indices_(
-      const std::vector<gp_Pnt2d>&            points,
-      const std::vector<std::pair<int, int>>& edge_indices,
-      const gp_Pln&                           plane,
-      Callback&&                              callback);
-
-  static GUI s_gui;
-  // std::unique_ptr<Occt_view> view;
-};
-
-GUI Sketch_test::s_gui;
-
-class Sketch_access
-{
- public:
-  static void add_edge_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b);
-  static void add_edge_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b, Sketch_op_recorder& rec);
-  static void add_edge_raw_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b);
-  static void update_faces_(Sketch& sketch);
-  static void add_arc_circle_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b, const gp_Pnt2d& pt_c);
-  static void add_arc_circle_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b, const gp_Pnt2d& pt_c,
-                              Sketch_op_recorder& rec);
-  static void get_originating_face_snp_pts_3d_(Sketch& sketch, std::vector<gp_Pnt>& out);
-
-  static const std::vector<Sketch_face_shp_ptr>& get_faces(const Sketch& sketch);
-  static const std::list<Sketch::Edge>&        get_edges(const Sketch& sketch);
-  static size_t                                 get_linear_edge_count(const Sketch& sketch);
-  static size_t                                 get_arc_internal_edge_count(const Sketch& sketch);
-  static size_t                                 length_dimension_count(const Sketch& sketch);
-  static size_t                                 length_dimension_node_lo(const Sketch& sketch, size_t index);
-  static size_t                                 length_dimension_node_hi(const Sketch& sketch, size_t index);
-  static void                                   set_entered_edge_len(Sketch& sketch, const gp_Dir2d& dir, double len);
-
-  /// Exact replay of a saved linear edge (with its pre-existing midpoint node index).
-  /// Used for regression tests of specific .ezy cases (e.g. bridge + hole topologies).
-  static void sketch_json_add_linear_edge_(Sketch& sketch, size_t idx_a, size_t idx_b, std::optional<size_t> idx_mid = std::nullopt);
-  static void set_operation_axis_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b);
-  static void remove_permanent_node_mark_(Sketch& sketch, size_t node_idx);
-};
-
-void Sketch_access::add_edge_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b)
-{
-  sketch.add_edge_(pt_a, pt_b);
-}
-
-void Sketch_access::add_edge_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b, Sketch_op_recorder& rec)
-{
-  sketch.add_edge_(pt_a, pt_b, rec);
-}
-
-void Sketch_access::add_edge_raw_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b)
-{
-  sketch.add_edge_raw_(pt_a, pt_b);
-}
-
-size_t Sketch_access::get_linear_edge_count(const Sketch& sketch)
-{
-  size_t count = 0;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-  {
-    if (sketch_edge_is_linear(edge))
-      ++count;
-  }
-  return count;
-}
-
-size_t Sketch_access::get_arc_internal_edge_count(const Sketch& sketch)
-{
-  size_t count = 0;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-  {
-    if (sketch_edge_is_arc(edge))
-      ++count;
-  }
-  return count;
-}
-
-const std::vector<Sketch_face_shp_ptr>& Sketch_access::get_faces(const Sketch& sketch)
-{
-  return sketch.m_topo.faces();
-}
-
-const std::list<Sketch::Edge>& Sketch_access::get_edges(const Sketch& sketch)
-{
-  return sketch.m_edges.edges();
-}
-
-size_t Sketch_access::length_dimension_count(const Sketch& sketch)
-{
-  return sketch.m_dims.dimensions().size();
-}
-
-size_t Sketch_access::length_dimension_node_lo(const Sketch& sketch, size_t index)
-{
-  return sketch.m_dims.dimensions()[index].node_idx_lo;
-}
-
-size_t Sketch_access::length_dimension_node_hi(const Sketch& sketch, size_t index)
-{
-  return sketch.m_dims.dimensions()[index].node_idx_hi;
-}
-
-void Sketch_access::set_entered_edge_len(Sketch& sketch, const gp_Dir2d& dir, double len)
-{
-  sketch.m_dims.entered_edge_len() = Sketch_dims::Edge_len{dir, len};
-}
-
-void Sketch_access::update_faces_(Sketch& sketch)
-{
-  sketch.update_faces_();
-}
-
-void Sketch_access::add_arc_circle_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b, const gp_Pnt2d& pt_c)
-{
-  sketch.add_arc_circle_(pt_a, pt_b, pt_c);
-}
-
-void Sketch_access::add_arc_circle_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b, const gp_Pnt2d& pt_c,
-                                    Sketch_op_recorder& rec)
-{
-  sketch.add_arc_circle_(pt_a, pt_b, pt_c, rec);
-}
-
-void GUI_access::set_view(GUI& gui, std::unique_ptr<Occt_view>& view)
-{
-  gui.m_view = std::move(view);
-}
-
-Occt_view& GUI_access::get_view(GUI& gui)
-{
-  return *gui.m_view;
-}
-
-std::string GUI_access::get_message(const GUI& gui)
-{
-  return gui.m_message;
-}
-
-void GUI_access::sketch_left_click(GUI& gui, const ScreenCoords& screen_coords)
-{
-  gui.sketch_left_click(screen_coords);
-}
-
-void GUI_access::mirror_selected_edges(GUI& gui)
-{
-  gui.mirror_selected_sketch_edges();
-}
-
-inline void Sketch_access::get_originating_face_snp_pts_3d_(Sketch& sketch, std::vector<gp_Pnt>& out)
-{
-  sketch.get_originating_face_snp_pts_3d_(out);
-}
-
-void Sketch_access::sketch_json_add_linear_edge_(Sketch& sketch, size_t idx_a, size_t idx_b, std::optional<size_t> idx_mid)
-{
-  sketch.sketch_json_add_linear_edge_(idx_a, idx_b, idx_mid);
-}
-
-void Sketch_access::set_operation_axis_(Sketch& sketch, const gp_Pnt2d& pt_a, const gp_Pnt2d& pt_b)
-{
-  sketch.sketch_json_set_operation_axis_(pt_a, pt_b);
-}
-
-void Sketch_access::remove_permanent_node_mark_(Sketch& sketch, size_t node_idx)
-{
-  Sketch_AIS_node_mark mark(sketch, node_idx, TopoDS_Shape());
-  sketch.remove_permanent_node_mark(mark);
-}
-
-class View_access
-{
- public:
-  static void set_view_plane(Occt_view& view, const gp_Pln& pln)
-  {
-    view.shp_extrude().set_curr_view_pln(pln);
-  }
-
-  static void set_headless(Occt_view& view, bool headless)
-  {
-    view.m_headless_view = headless;
-  }
-};
 
 // Test basic sketch creation and initialization
 TEST_F(Sketch_test, CreateSketch)
@@ -302,6 +58,8 @@ TEST_F(Sketch_test, NodeManagement)
 }
 
 // Test edge creation and management
+
+// Test edge creation and management
 TEST_F(Sketch_test, EdgeManagement)
 {
   gp_Pln default_plane(gp::Origin(), gp::DZ());
@@ -321,6 +79,10 @@ TEST_F(Sketch_test, EdgeManagement)
   // Origin node + 2 endpoints + 1 midpoint (midpoints enabled in fixture).
   EXPECT_EQ(sketch.get_nodes().size(), 4u);
 }
+
+// Test that adding a single edge via add_edge_ creates exactly three nodes:
+// the two endpoints (user nodes) + one automatically created midpoint node for snapping.
+// Verifies positions and the midpoint flag.
 
 // Test that adding a single edge via add_edge_ creates exactly three nodes:
 // the two endpoints (user nodes) + one automatically created midpoint node for snapping.
@@ -356,6 +118,10 @@ TEST_F(Sketch_test, AddSingleEdge_CreatesThreeCorrectNodes)
   EXPECT_TRUE(nodes[2].IsEqual(expected_mid, Precision::Confusion())) << "Node 2 should be the exact midpoint";
   EXPECT_TRUE(nodes[2].midpoint) << "The auto-created center node must be marked as midpoint for snapping";
 }
+
+// Test the new "add midpoints to linear edges" option (default off).
+// Verifies that linear edges do not get auto midpoints by default,
+// and that the static setter controls it for add_edge_ (used by line/multi-line tools).
 
 // Test the new "add midpoints to linear edges" option (default off).
 // Verifies that linear edges do not get auto midpoints by default,
@@ -467,30 +233,6 @@ TEST_F(Sketch_test, AddArc_MidpointOption)
 // automatically-created midpoint results in each being split into two sub-edges, for a total of 4 edges.
 // This exercises the "split existing intersecting/touching edges on add" logic (and the corresponding
 // subdivision of the new edge) in the non-midpoint-crossing case.
-TEST_F(Sketch_test, AddTwoCrossingEdges_NotAtMidpoints_ProducesFourEdges)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // First edge: horizontal (0,0) to (10,0). Its auto mid will be at (5,0).
-  gp_Pnt2d p1(0.0, 0.0);
-  gp_Pnt2d p2(10.0, 0.0);
-  Sketch_access::add_edge_(sketch, p1, p2);
-
-  // Second edge: vertical crossing the first at (3,0).
-  // - For the horizontal, 3 != 5, so not at its midpoint.
-  // - For the vertical from (3,-2) to (3,4), its auto mid would be at (3,1); 0 != 1 so not at midpoint either.
-  // The interior intersection (away from mids) should cause:
-  //   * the existing horizontal to be split at (3,0) into two edges (each with own mid)
-  //   * the new vertical to be subdivided at (3,0) into two raw sub-edges (each with own mid)
-  // Result: 4 linear edges in the sketch.
-  gp_Pnt2d p3(3.0, -2.0);
-  gp_Pnt2d p4(3.0, 4.0);
-  Sketch_access::add_edge_(sketch, p3, p4);
-
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 4)
-      << "Two edges crossing interiorly (but not at midpoints) must each be split, producing four edges total";
-}
 
 // Test add line edge with "place from center": Tab length is the full edge span.
 TEST_F(Sketch_test, AddEdgeFromCenter_SecondClickFullSpan)
@@ -572,346 +314,6 @@ TEST_F(Sketch_test, AddEdgeFromCenter_DimensionUsesFullLength)
 // *exactly through the midpoint* of the first (and not the midpoint of the second).
 // The split logic must still split the horizontal at its (former) mid node and subdivide the
 // vertical, resulting in four edges total. (Currently observed to produce only three.)
-TEST_F(Sketch_test, AddTwoCrossingEdges_ThroughMidpoint_ProducesFourEdges)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // Simulate GUI-style edge addition (exercises finalize_edges_ path, including the pre-pass
-  // for splitting olds at inters and the post-append re-subdivide for news). Horizontal first.
-  gui().set_mode(Mode::Sketch_add_edge);
-
-  // Horizontal: (0,0) to (10,0). Uses fake screen coords (world values) like other creation tests.
-  // After finalize: 1 edge + its mid at (5,0).
-  ScreenCoords sc1(dvec2(0.0, 0.0));
-  sketch.add_sketch_pt(sc1);
-  ScreenCoords sc2(dvec2(10.0, 0.0));
-  sketch.add_sketch_pt(sc2);
-  sketch.finalize_elm();
-
-  // Now the vertical, passing exactly through the horizontal's midpoint (5,0).
-  // Vertical from (5,-5) to (5,10): the cross (5,0) is interior to horiz (its mid) and
-  // interior to this vert (its mid would be (5,2.5), 0 != 2.5).
-  // With the fix, horiz will be split at mid + vert will be subdivided at cross --> 4 edges.
-  gui().set_mode(Mode::Sketch_add_edge);
-  ScreenCoords sc3(dvec2(5.0, -5.0));
-  sketch.add_sketch_pt(sc3);
-  ScreenCoords sc4(dvec2(5.0, 10.0));
-  sketch.add_sketch_pt(sc4);
-  sketch.finalize_elm();
-
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 4)
-      << "Vertical through existing horizontal's midpoint (GUI finalize path) must still cause splits on both, yielding four edges";
-}
-
-TEST_F(Sketch_test, Undo_single_arc_via_recorder)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  const gp_Pnt2d start(0.0, 0.0);
-  const gp_Pnt2d end(10.0, 0.0);
-  const gp_Pnt2d arc_mid(5.0, 5.0);
-
-  // Arc Segment tool: recorder starts on the third click, after all pick nodes exist.
-  gui().set_mode(Mode::Sketch_add_seg_circle_arc);
-  sketch.add_sketch_pt(ScreenCoords(dvec2(start.X(), start.Y())));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(end.X(), end.Y())));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(arc_mid.X(), arc_mid.Y())));
-
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(sketch), 1);
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(sketch), 0)
-      << "Undo should remove the arc edge";
-
-  EXPECT_TRUE(view().redo());
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(sketch), 1)
-      << "Redo should restore the arc";
-}
-
-TEST_F(Sketch_test, Undo_circle_via_recorder)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_add_circle);
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 0.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(5.0, 0.0)));
-  sketch.finalize_elm();
-
-  auto count_permanent_nodes = [](Sketch& s) {
-    size_t n = 0;
-    for (const auto& node : s.get_nodes())
-      if (node.permanent && !node.deleted)
-        ++n;
-    return n;
-  };
-
-  const size_t edges_after_create = Sketch_access::get_arc_internal_edge_count(sketch);
-  const size_t perm_after_create  = count_permanent_nodes(sketch);
-  EXPECT_EQ(edges_after_create, 2u);
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(sketch), 0u)
-      << "Undo should remove both semicircles (two arc edges)";
-
-  EXPECT_TRUE(view().redo());
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(sketch), edges_after_create)
-      << "Redo should restore the circle";
-  EXPECT_EQ(count_permanent_nodes(sketch), perm_after_create)
-      << "Redo should not add extra permanent nodes";
-}
-
-TEST_F(Sketch_test, Undo_two_crossing_edges_via_finalize)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  {
-    Sketch_op_recorder rec(view(), sketch);
-    Sketch_access::add_edge_(sketch, gp_Pnt2d(0.0, 0.0), gp_Pnt2d(10.0, 0.0), rec);
-    rec.commit();
-  }
-
-  {
-    Sketch_op_recorder rec(view(), sketch);
-    Sketch_access::add_edge_(sketch, gp_Pnt2d(3.0, -2.0), gp_Pnt2d(3.0, 4.0), rec);
-    rec.commit();
-  }
-
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 4);
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 1)
-      << "Undo of the second edge should remove it and restore the unsplit first edge";
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 0)
-      << "Undo of the first edge should remove it";
-}
-
-TEST_F(Sketch_test, Redo_two_crossing_edges_keeps_intersection_node_non_permanent)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  auto count_permanent_nodes = [](Sketch& s) {
-    size_t n = 0;
-    for (const auto& node : s.get_nodes())
-      if (node.permanent && !node.deleted)
-        ++n;
-    return n;
-  };
-
-  auto find_live_node_at = [](Sketch& s, const gp_Pnt2d& pt) -> std::optional<size_t> {
-    for (size_t i = 0, n = s.get_nodes().size(); i < n; ++i)
-    {
-      const auto& node = s.get_nodes()[i];
-      if (node.deleted)
-        continue;
-
-      if (node.SquareDistance(pt) <= Precision::SquareConfusion())
-        return i;
-    }
-    return std::nullopt;
-  };
-
-  {
-    Sketch_op_recorder rec(view(), sketch);
-    Sketch_access::add_edge_(sketch, gp_Pnt2d(0.0, 0.0), gp_Pnt2d(10.0, 0.0), rec);
-    rec.commit();
-  }
-
-  {
-    Sketch_op_recorder rec(view(), sketch);
-    Sketch_access::add_edge_(sketch, gp_Pnt2d(3.0, -2.0), gp_Pnt2d(3.0, 4.0), rec);
-    rec.commit();
-  }
-
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 4);
-  const size_t perm_after_create = count_permanent_nodes(sketch);
-
-  const std::optional<size_t> intersection_before = find_live_node_at(sketch, gp_Pnt2d(3.0, 0.0));
-  ASSERT_TRUE(intersection_before.has_value()) << "Crossing should create an interior node";
-  EXPECT_FALSE(sketch.get_nodes()[*intersection_before].permanent)
-      << "Auto-split intersection nodes should not be permanent";
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_TRUE(view().undo());
-  EXPECT_TRUE(view().redo());
-  EXPECT_TRUE(view().redo());
-
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 4);
-  EXPECT_EQ(count_permanent_nodes(sketch), perm_after_create) << "Redo should not add permanent nodes";
-
-  const std::optional<size_t> intersection_after = find_live_node_at(sketch, gp_Pnt2d(3.0, 0.0));
-  ASSERT_TRUE(intersection_after.has_value()) << "Crossing intersection should be restored";
-  EXPECT_FALSE(sketch.get_nodes()[*intersection_after].permanent)
-      << "Redo must not promote the intersection node to permanent";
-}
-
-TEST_F(Sketch_test, Undo_extrude_snapshot_keeps_single_sketch)
-{
-  Sketch::Sketch_ptr sk = view().curr_sketch_shared();
-  const size_t sk_id = sk->get_id();
-
-  {
-    Sketch_op_recorder rec(view(), *sk);
-    const std::array<gp_Pnt2d, 4> corners = square_corners(gp_Pnt2d(0.0, 0.0), gp_Pnt2d(10.0, 0.0));
-    Sketch_access::add_edge_(*sk, corners[0], corners[1], rec);
-    Sketch_access::add_edge_(*sk, corners[1], corners[2], rec);
-    Sketch_access::add_edge_(*sk, corners[2], corners[3], rec);
-    Sketch_access::add_edge_(*sk, corners[3], corners[0], rec);
-    rec.commit();
-  }
-
-  EXPECT_EQ(sk->edge_count(), 4u);
-  EXPECT_EQ(view().get_sketches().size(), 1u);
-
-  const nlohmann::json pre_extrude = nlohmann::json::parse(view().to_json());
-  ASSERT_EQ(pre_extrude["sketches"].size(), 1u);
-
-  view().push_undo_snapshot();
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_EQ(view().get_sketches().size(), 1u) << "JSON undo after extrude must not add a sketch";
-
-  Sketch::Sketch_ptr restored;
-  for (const Sketch::Sketch_ptr& s : view().get_sketches())
-    if (s->get_id() == sk_id)
-      restored = s;
-
-  ASSERT_TRUE(restored) << "Restored sketch should keep the same stable id";
-  EXPECT_EQ(restored->edge_count(), 4u);
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_EQ(restored->edge_count(), 0u) << "Second undo should remove the square from the same sketch";
-}
-
-TEST_F(Sketch_test, Undo_delta_resolves_by_sketch_id_when_names_duplicate)
-{
-  Sketch::Sketch_ptr sk1 = view().curr_sketch_shared();
-  Sketch::Sketch_ptr sk2 = std::make_shared<Sketch>("Other", view(), gp_Pln(gp::Origin(), gp::DX()));
-  view().get_sketches().push_back(sk2);
-  sk1->set_name("Dup");
-  sk2->set_name("Dup");
-
-  {
-    Sketch_op_recorder rec(view(), *sk2);
-    Sketch_access::add_edge_(*sk2, gp_Pnt2d(0.0, 0.0), gp_Pnt2d(5.0, 0.0), rec);
-    rec.commit();
-  }
-
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(*sk1), 0u);
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(*sk2), 1u);
-
-  EXPECT_TRUE(view().undo());
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(*sk1), 0u);
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(*sk2), 0u)
-      << "Undo must target the sketch that was edited, not the first same-named sketch";
-}
-
-TEST_F(Sketch_test, Undo_circle_axis_then_revolve_snapshot_three_undos)
-{
-  Sketch::Sketch_ptr sk = view().curr_sketch_shared();
-  const size_t         sk_id = sk->get_id();
-
-  {
-    Sketch_op_recorder rec(view(), *sk);
-    const gp_Pnt2d                 center(0.0, 0.0);
-    const gp_Pnt2d                 edge_pt(5.0, 0.0);
-    const std::array<gp_Pnt2d, 4> points = xy_stencil_pnts(center, edge_pt);
-    Sketch_access::add_arc_circle_(*sk, points[0], points[2], points[1], rec);
-    Sketch_access::add_arc_circle_(*sk, points[0], points[3], points[1], rec);
-    rec.commit();
-  }
-
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(*sk), 2u);
-
-  {
-    Sketch_op_recorder rec(view(), *sk);
-    const gp_Pnt2d axis_a(0.0, -10.0);
-    const gp_Pnt2d axis_b(10.0, -10.0);
-    Sketch_access::set_operation_axis_(*sk, axis_a, axis_b);
-    rec.note_curr_operation_axis(axis_a, axis_b);
-    rec.commit();
-  }
-  EXPECT_TRUE(sk->has_operation_axis());
-
-  view().push_undo_snapshot();
-
-  auto find_sk = [&]() -> Sketch::Sketch_ptr
-  {
-    for (const Sketch::Sketch_ptr& s : view().get_sketches())
-      if (s->get_id() == sk_id)
-        return s;
-    return nullptr;
-  };
-
-  EXPECT_TRUE(view().undo());
-  sk = find_sk();
-  ASSERT_TRUE(sk);
-  EXPECT_EQ(view().get_sketches().size(), 1u);
-  EXPECT_TRUE(sk->has_operation_axis());
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(*sk), 2u);
-
-  EXPECT_TRUE(view().undo());
-  sk = find_sk();
-  ASSERT_TRUE(sk);
-  EXPECT_FALSE(sk->has_operation_axis());
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(*sk), 2u);
-
-  EXPECT_TRUE(view().undo());
-  sk = find_sk();
-  ASSERT_TRUE(sk);
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(*sk), 0u);
-}
-
-// Test T-junction case (one edge's endpoint touches the interior of another).
-// Initially two edges are added; the touched edge is split at the junction point,
-// resulting in three edges total. Test both addition orders.
-TEST_F(Sketch_test, AddTwoEdges_TJunction_ProducesThreeEdges)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-
-  // Horizontal first, then vertical T-junction (attach point is interior to horizontal, not its midpoint).
-  //   0--------10   (horizontal)
-  //          |
-  //          |      (vertical attaches at x=3)
-  //          |
-  {
-    Sketch sketch("TestSketch", view(), default_plane);
-
-    gp_Pnt2d h1(0.0, 0.0);
-    gp_Pnt2d h2(10.0, 0.0);
-    Sketch_access::add_edge_(sketch, h1, h2);
-
-    gp_Pnt2d v1(3.0, 0.0);  // attaches to interior of horizontal (3 != 5 midpoint)
-    gp_Pnt2d v2(3.0, -5.0);
-    Sketch_access::add_edge_(sketch, v1, v2);
-
-    EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 3)
-        << "T-junction (horiz first): existing edge split at interior attach point -> 3 edges total";
-  }
-
-  // Vertical first, then horizontal (the horizontal's interior will touch the vertical's endpoint).
-  // Same geometry, different add order.
-  {
-    Sketch sketch("TestSketch", view(), default_plane);
-
-    gp_Pnt2d v1(3.0, 0.0);
-    gp_Pnt2d v2(3.0, -5.0);
-    Sketch_access::add_edge_(sketch, v1, v2);
-
-    gp_Pnt2d h1(0.0, 0.0);
-    gp_Pnt2d h2(10.0, 0.0);
-    Sketch_access::add_edge_(sketch, h1, h2);
-
-    EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 3)
-        << "T-junction (vert first): new edge split at interior attach point on it -> 3 edges total";
-  }
-}
 
 // Test visibility settings
 TEST_F(Sketch_test, VisibilitySettings)
@@ -935,296 +337,6 @@ TEST_F(Sketch_test, VisibilitySettings)
 // Test for splitting a square with a vertical edge at the midpoints (T-junction at mid, not crossing).
 // Result must be two proper rectangles in *both* addition orders.
 // (Image shows the geometry: square split vertically down the middle into two rects.)
-TEST_F(Sketch_test, AddSquareThenMidEdge_ProducesTwoRectangles_BothOrders)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-
-  // Square: side 10 (will be split into two 5x10 rectangles)
-  gp_Pnt2d bl(0.0, 0.0);
-  gp_Pnt2d br(10.0, 0.0);
-  gp_Pnt2d tr(10.0, 10.0);
-  gp_Pnt2d tl(0.0, 10.0);
-
-  // Mid vertical edge
-  gp_Pnt2d m_bottom(5.0, 0.0);
-  gp_Pnt2d m_top(5.0, 10.0);
-
-  // Order 1: square (4 edges) first, then the mid edge
-  {
-    Sketch sketch("square_then_mid", view(), default_plane);
-
-    // Add square (order doesn't matter much for the square itself, but we add sequentially)
-    Sketch_access::add_edge_(sketch, bl, br);
-    Sketch_access::add_edge_(sketch, br, tr);
-    Sketch_access::add_edge_(sketch, tr, tl);
-    Sketch_access::add_edge_(sketch, tl, bl);
-
-    // Add the splitting edge at midpoints
-    Sketch_access::add_edge_(sketch, m_bottom, m_top);
-
-    Sketch_access::update_faces_(sketch);
-    const auto& faces = Sketch_access::get_faces(sketch);
-
-    EXPECT_EQ(faces.size(), 2) << "Square + mid vertical must produce exactly two faces";
-
-    // Check both faces are proper 5x10 rectangles (not squares)
-    bool found_left = false, found_right = false;
-    for (const auto& f : faces)
-    {
-      EXPECT_EQ(f->Shape().ShapeType(), TopAbs_FACE);
-      auto poly = to_boost(f->Shape(), default_plane);
-      EXPECT_TRUE(is_clockwise(poly.outer()));
-      EXPECT_EQ(poly.outer().size(), 5); // closed ring
-
-      // Compute bbox deltas
-      double minx = 1e9, maxx = -1e9, miny = 1e9, maxy = -1e9;
-      for (const auto& pt : poly.outer())
-      {
-        minx = std::min(minx, pt.x());
-        maxx = std::max(maxx, pt.x());
-        miny = std::min(miny, pt.y());
-        maxy = std::max(maxy, pt.y());
-      }
-      double dx = maxx - minx;
-      double dy = maxy - miny;
-
-      if (std::abs(dx - 5.0) < 1e-6 && std::abs(dy - 10.0) < 1e-6)
-      {
-        // left or right rect
-        if (minx < 2.5) found_left = true;
-        else found_right = true;
-      }
-      else
-      {
-        ADD_FAILURE() << "Face is not a 5x10 rectangle (dx=" << dx << " dy=" << dy << ")";
-      }
-    }
-    EXPECT_TRUE(found_left) << "Missing left 5x10 rect";
-    EXPECT_TRUE(found_right) << "Missing right 5x10 rect";
-  }
-
-  // Order 2: mid edge first, then the square edges
-  {
-    Sketch sketch("mid_then_square", view(), default_plane);
-
-    // Add mid vertical first
-    Sketch_access::add_edge_(sketch, m_bottom, m_top);
-
-    // Add square (the square edges will attach to the existing mid edge)
-    Sketch_access::add_edge_(sketch, bl, br);
-    Sketch_access::add_edge_(sketch, br, tr);
-    Sketch_access::add_edge_(sketch, tr, tl);
-    Sketch_access::add_edge_(sketch, tl, bl);
-
-    Sketch_access::update_faces_(sketch);
-    const auto& faces = Sketch_access::get_faces(sketch);
-
-    EXPECT_EQ(faces.size(), 2) << "Mid vertical + square must produce exactly two faces";
-
-    bool found_left = false, found_right = false;
-    for (const auto& f : faces)
-    {
-      EXPECT_EQ(f->Shape().ShapeType(), TopAbs_FACE);
-      auto poly = to_boost(f->Shape(), default_plane);
-      EXPECT_TRUE(is_clockwise(poly.outer()));
-      EXPECT_EQ(poly.outer().size(), 5);
-
-      double minx = 1e9, maxx = -1e9, miny = 1e9, maxy = -1e9;
-      for (const auto& pt : poly.outer())
-      {
-        minx = std::min(minx, pt.x());
-        maxx = std::max(maxx, pt.x());
-        miny = std::min(miny, pt.y());
-        maxy = std::max(maxy, pt.y());
-      }
-      double dx = maxx - minx;
-      double dy = maxy - miny;
-
-      if (std::abs(dx - 5.0) < 1e-6 && std::abs(dy - 10.0) < 1e-6)
-      {
-        if (minx < 2.5) found_left = true;
-        else found_right = true;
-      }
-      else
-      {
-        ADD_FAILURE() << "Face is not a 5x10 rectangle (dx=" << dx << " dy=" << dy << ")";
-      }
-    }
-    EXPECT_TRUE(found_left) << "Missing left 5x10 rect (mid first)";
-    EXPECT_TRUE(found_right) << "Missing right 5x10 rect (mid first)";
-  }
-}
-
-// GUI path coverage for the mid edge: the vertical mid is added via the finalize_edges_
-// (line string tmp, pre-pass batch inters to split olds, append, post re-subdivide news,
-// plus mid snap handling). Exercise both draw directions for the vertical (lower midpoint
-// first then top; top first then lower). Must still produce exactly two 5x10 rect faces
-// (not squares) in either case. This addresses order/direction sensitivity that can appear
-// only in the GUI finalize path vs controlled direct add_edge_.
-TEST_F(Sketch_test, AddSquareThenMidEdge_GUIPath_BothDrawDirections_ProducesTwoRectangles)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-
-  // Square: side 10 (will be split into two 5x10 rectangles)
-  gp_Pnt2d bl(0.0, 0.0);
-  gp_Pnt2d br(10.0, 0.0);
-  gp_Pnt2d tr(10.0, 10.0);
-  gp_Pnt2d tl(0.0, 10.0);
-
-  // Mid vertical edge
-  gp_Pnt2d m_bottom(5.0, 0.0);
-  gp_Pnt2d m_top(5.0, 10.0);
-
-  // Draw square via direct (its finalize uses direct add_edge_ for the 4 sides), then
-  // add the mid vertical using the GUI single-edge finalize path, lower to top.
-  {
-    Sketch sketch("gui_square_then_mid_lower_first", view(), default_plane);
-
-    Sketch_access::add_edge_(sketch, bl, br);
-    Sketch_access::add_edge_(sketch, br, tr);
-    Sketch_access::add_edge_(sketch, tr, tl);
-    Sketch_access::add_edge_(sketch, tl, bl);
-
-    // Simulate GUI draw of vertical: click lower mid first, then top.
-    gui().set_mode(Mode::Sketch_add_edge);
-    sketch.add_sketch_pt(ScreenCoords(dvec2(m_bottom.X(), m_bottom.Y())));
-    sketch.add_sketch_pt(ScreenCoords(dvec2(m_top.X(), m_top.Y())));
-    sketch.finalize_elm();
-
-    Sketch_access::update_faces_(sketch);
-    const auto& faces = Sketch_access::get_faces(sketch);
-
-    EXPECT_EQ(faces.size(), 2) << "GUI mid vertical (lower first) must produce exactly two faces";
-
-    bool found_left = false, found_right = false;
-    for (const auto& f : faces)
-    {
-      EXPECT_EQ(f->Shape().ShapeType(), TopAbs_FACE);
-      auto poly = to_boost(f->Shape(), default_plane);
-      EXPECT_TRUE(is_clockwise(poly.outer()));
-      EXPECT_EQ(poly.outer().size(), 5);
-
-      double minx = 1e9, maxx = -1e9, miny = 1e9, maxy = -1e9;
-      for (const auto& pt : poly.outer())
-      {
-        minx = std::min(minx, pt.x());
-        maxx = std::max(maxx, pt.x());
-        miny = std::min(miny, pt.y());
-        maxy = std::max(maxy, pt.y());
-      }
-      double dx = maxx - minx;
-      double dy = maxy - miny;
-
-      if (std::abs(dx - 5.0) < 1e-6 && std::abs(dy - 10.0) < 1e-6)
-      {
-        if (minx < 2.5) found_left = true;
-        else found_right = true;
-      }
-      else
-      {
-        ADD_FAILURE() << "GUI (lower first): Face is not a 5x10 rectangle (dx=" << dx << " dy=" << dy << ")";
-      }
-    }
-    EXPECT_TRUE(found_left) << "GUI (lower first): Missing left 5x10 rect";
-    EXPECT_TRUE(found_right) << "GUI (lower first): Missing right 5x10 rect";
-  }
-
-  // Same square, but draw the mid vertical top to bottom (reverse dir).
-  {
-    Sketch sketch("gui_square_then_mid_top_first", view(), default_plane);
-
-    Sketch_access::add_edge_(sketch, bl, br);
-    Sketch_access::add_edge_(sketch, br, tr);
-    Sketch_access::add_edge_(sketch, tr, tl);
-    Sketch_access::add_edge_(sketch, tl, bl);
-
-    gui().set_mode(Mode::Sketch_add_edge);
-    sketch.add_sketch_pt(ScreenCoords(dvec2(m_top.X(), m_top.Y())));
-    sketch.add_sketch_pt(ScreenCoords(dvec2(m_bottom.X(), m_bottom.Y())));
-    sketch.finalize_elm();
-
-    Sketch_access::update_faces_(sketch);
-    const auto& faces = Sketch_access::get_faces(sketch);
-
-    EXPECT_EQ(faces.size(), 2) << "GUI mid vertical (top first) must produce exactly two faces";
-
-    bool found_left = false, found_right = false;
-    for (const auto& f : faces)
-    {
-      EXPECT_EQ(f->Shape().ShapeType(), TopAbs_FACE);
-      auto poly = to_boost(f->Shape(), default_plane);
-      EXPECT_TRUE(is_clockwise(poly.outer()));
-      EXPECT_EQ(poly.outer().size(), 5);
-
-      double minx = 1e9, maxx = -1e9, miny = 1e9, maxy = -1e9;
-      for (const auto& pt : poly.outer())
-      {
-        minx = std::min(minx, pt.x());
-        maxx = std::max(maxx, pt.x());
-        miny = std::min(miny, pt.y());
-        maxy = std::max(maxy, pt.y());
-      }
-      double dx = maxx - minx;
-      double dy = maxy - miny;
-
-      if (std::abs(dx - 5.0) < 1e-6 && std::abs(dy - 10.0) < 1e-6)
-      {
-        if (minx < 2.5) found_left = true;
-        else found_right = true;
-      }
-      else
-      {
-        ADD_FAILURE() << "GUI (top first): Face is not a 5x10 rectangle (dx=" << dx << " dy=" << dy << ")";
-      }
-    }
-    EXPECT_TRUE(found_left) << "GUI (top first): Missing left 5x10 rect";
-    EXPECT_TRUE(found_right) << "GUI (top first): Missing right 5x10 rect";
-  }
-}
-
-// Two squares via the GUI square tool (finalize_square_ path): a 2x2 square (radius/half-side 1)
-// plus a 1x1 square placed with its center on the top-right corner of the first. With midpoints
-// off, the internal edges of the second square should split the first into three faces (not two).
-TEST_F(Sketch_test, TwoSquares_TopRightCornerOverlap_GUIPath_ProducesThreeFaces)
-{
-  struct Snap_dist_guard
-  {
-    const double prev;
-    Snap_dist_guard()
-        : prev(Sketch_nodes::get_snap_dist())
-    {
-      // Headless snap radius is ~35 plane units; disable so a 1" half-side click is not pulled to center.
-      Sketch_nodes::set_snap_dist(0.5);
-    }
-    ~Snap_dist_guard() { Sketch_nodes::set_snap_dist(prev); }
-  } snap_guard;
-
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("two_squares_top_right", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_add_square);
-  Sketch::set_add_mid_pt_edges(false);
-
-  // First square: center (0,0), edge midpoint at (1,0) => half-side 1 (2x2 square).
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 0.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(1.0, 0.0)));
-  sketch.finalize_elm();
-
-  // Second 1" square: center at top-right corner (1,1) of the first, half-side 0.5.
-  sketch.add_sketch_pt(ScreenCoords(dvec2(1.0, 1.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(1.5, 1.0)));
-  sketch.finalize_elm();
-
-  for (const auto& n : sketch.get_nodes())
-  {
-    if (!n.deleted)
-      EXPECT_FALSE(n.midpoint) << "Square edges should not have auto midpoint nodes";
-  }
-
-  Sketch_access::update_faces_(sketch);
-  const auto& faces = Sketch_access::get_faces(sketch);
-  EXPECT_EQ(faces.size(), 3) << "Overlapping corner squares should produce three faces";
-}
 
 // Test edge style settings
 TEST_F(Sketch_test, EdgeStyleSettings)
@@ -1246,6 +358,8 @@ TEST_F(Sketch_test, EdgeStyleSettings)
   sketch.set_edge_style(Sketch::Edge_style::Background);
   // Note: We can't directly test the visual style, but we can verify the setting was applied
 }
+
+// Test square creation
 
 // Test square creation
 TEST_F(Sketch_test, CreateSquare)
@@ -1335,6 +449,8 @@ TEST_F(Sketch_test, CreateSquare)
 }
 
 // Test circle creation
+
+// Test circle creation
 TEST_F(Sketch_test, CreateCircle)
 {
   gp_Pln default_plane(gp::Origin(), gp::DZ());
@@ -1374,440 +490,6 @@ TEST_F(Sketch_test, CreateCircle)
 }
 
 // Test case 1: Simple rectangle face
-TEST_F(Sketch_test, UpdateFaces_SimpleRectangle)
-{
-  // Define points
-  std::vector<gp_Pnt2d> points = {
-      { 0,  0}, // 0
-      {10,  0}, // 1
-      {10, 10}, // 2
-      { 0, 10}  // 3
-  };
-
-  // Define edges as pairs of indices
-  std::vector<std::pair<int, int>> edge_indices = {
-      {0, 1},
-      {1, 2},
-      {2, 3},
-      {3, 0}
-  };
-
-  auto check_sketch = [&](Sketch& sketch)
-  {
-    Sketch_access::update_faces_(sketch);
-    const auto& faces = Sketch_access::get_faces(sketch);
-    EXPECT_EQ(faces.size(), 1);
-    // Verify face properties
-    const auto& face = faces[0];
-    EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE);
-
-    ezy_geom::polygon_2d poly = to_boost(face->Shape(), sketch.get_plane());
-    // Orientation of the produced poly can vary with traversal seeds/adj order and to_boost wire handling;
-    // the important is that a single valid face of correct area is always produced regardless of add order.
-    EXPECT_TRUE(ezy_geom::is_valid(poly));
-    EXPECT_EQ(poly.outer().size(), 5);
-    double area = ezy_geom::area(poly);
-    EXPECT_NEAR(area, 100.0, 1e-6);
-  };
-
-  for_all_edge_permutations_(points, edge_indices, gp_Pln(gp::Origin(), gp::DZ()), check_sketch);
-}
-
-// Test case 2: Face with hole
-TEST_F(Sketch_test, UpdateFaces_FaceWithHole)
-{
-  Sketch sketch("test_sketch", view(), gp_Pln(gp::Origin(), gp::DZ()));
-
-  // Define points
-  std::vector<gp_Pnt2d> points = {
-      { 0,  0}, // 0
-      {20,  0}, // 1
-      {20, 20}, // 2
-      { 0, 20}, // 3
-      { 5,  5}, // 4
-      {15,  5}, // 5
-      {15, 15}, // 6
-      { 5, 15}  // 7
-  };
-
-  // Define edges as pairs of indices
-  std::vector<std::pair<int, int>> edge_indices = {
-      {0, 1},
-      {1, 2},
-      {2, 3},
-      {3, 0}, // Outer rectangle
-      {4, 5},
-      {5, 6},
-      {6, 7},
-      {7, 4}  // Inner rectangle (hole)
-  };
-
-  auto check_sketch = [&](Sketch& sketch)
-  {
-    Sketch_access::update_faces_(sketch);
-    const auto& faces = Sketch_access::get_faces(sketch);
-    EXPECT_EQ(faces.size(), 2);
-    ezy_geom::polygon_2d face_with_hole;
-    ezy_geom::polygon_2d hole_face;
-    for (auto& face : faces)
-    {
-      // Verify face properties
-      EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE);
-      ezy_geom::polygon_2d poly = to_boost(face->Shape(), sketch.get_plane());
-      EXPECT_TRUE(is_clockwise(poly.outer()));
-      EXPECT_FALSE(poly.outer().empty());
-      if (poly.inners().size())
-      {
-        EXPECT_TRUE(face_with_hole.outer().empty());
-        std::swap(face_with_hole, poly);
-      }
-      else
-      {
-        EXPECT_TRUE(hole_face.outer().empty());
-        std::swap(hole_face, poly);
-      }
-    }
-    // to_boost produces closed rings (first==last), so a 4-corner rect ring has size 5.
-    // Verify structure: one face is the separate "hole" rect (no inners), the other is outer with 1 inner ring.
-    EXPECT_EQ(hole_face.outer().size(), 5);
-    EXPECT_EQ(hole_face.inners().size(), 0);
-    EXPECT_EQ(face_with_hole.outer().size(), 5);
-    EXPECT_EQ(face_with_hole.inners().size(), 1);
-  };
-
-  add_edges_from_indices_(points, edge_indices, gp_Pln(gp::Origin(), gp::DZ()), check_sketch);
-}
-
-// Test case 3: Multiple faces
-TEST_F(Sketch_test, UpdateFaces_MultipleFaces)
-{
-  Sketch sketch("test_sketch", view(), gp_Pln(gp::Origin(), gp::DZ()));
-
-  // Create two separate rectangles
-  // Rectangle 1
-  gp_Pnt2d p1(0, 0);
-  gp_Pnt2d p2(10, 0);
-  gp_Pnt2d p3(10, 10);
-  gp_Pnt2d p4(0, 10);
-
-  // Rectangle 2
-  gp_Pnt2d p5(20, 0);
-  gp_Pnt2d p6(30, 0);
-  gp_Pnt2d p7(30, 10);
-  gp_Pnt2d p8(20, 10);
-
-  // Add edges for first rectangle
-  Sketch_access::add_edge_(sketch, p1, p2);
-  Sketch_access::add_edge_(sketch, p2, p3);
-  Sketch_access::add_edge_(sketch, p3, p4);
-  Sketch_access::add_edge_(sketch, p4, p1);
-
-  // Add edges for second rectangle
-  Sketch_access::add_edge_(sketch, p5, p6);
-  Sketch_access::add_edge_(sketch, p6, p7);
-  Sketch_access::add_edge_(sketch, p7, p8);
-  Sketch_access::add_edge_(sketch, p8, p5);
-
-  Sketch_access::update_faces_(sketch);
-
-  // Verify two faces were created
-  const auto& faces = Sketch_access::get_faces(sketch);
-  EXPECT_EQ(faces.size(), 2);
-
-  // Verify each face's properties
-  for (const auto& face : faces)
-  {
-    EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE);
-    EXPECT_EQ(face->verts_3d.size(), 5);
-  }
-}
-
-// Test case 4: Invalid face (non-closed shape)
-TEST_F(Sketch_test, UpdateFaces_InvalidFace)
-{
-  Sketch sketch("test_sketch", view(), gp_Pln(gp::Origin(), gp::DZ()));
-
-  // Add three edges that don't form a closed shape
-  gp_Pnt2d p1(0, 0);
-  gp_Pnt2d p2(10, 0);
-  gp_Pnt2d p3(10, 10);
-
-  Sketch_access::add_edge_(sketch, p1, p2);
-  Sketch_access::add_edge_(sketch, p2, p3);
-
-  Sketch_access::update_faces_(sketch);
-
-  // Verify no faces were created
-  const auto& faces = Sketch_access::get_faces(sketch);
-  EXPECT_EQ(faces.size(), 0);
-}
-
-// Test case 5: Face with arc edges
-TEST_F(Sketch_test, UpdateFaces_FaceWithArcs)
-{
-  Sketch sketch("test_sketch", view(), gp_Pln(gp::Origin(), gp::DZ()));
-
-  // Create a face with straight edges and arc edges
-  gp_Pnt2d p1(0, 0);
-  gp_Pnt2d p2(10, 0);
-  gp_Pnt2d p3(10, 10);
-  gp_Pnt2d p4(0, 10);
-  gp_Pnt2d p5(5, 5);  // Center point for arc
-
-  // Add straight edges
-  Sketch_access::add_edge_(sketch, p1, p2);
-  Sketch_access::add_edge_(sketch, p2, p3);
-  Sketch_access::add_edge_(sketch, p3, p4);
-
-  // Add arc edge using Sketch_access
-  Sketch_access::add_arc_circle_(sketch, p4, p5, p1);
-
-  Sketch_access::update_faces_(sketch);
-
-  // Verify one face was created
-  const auto& faces = Sketch_access::get_faces(sketch);
-  EXPECT_EQ(faces.size(), 1);
-
-  // Verify face properties
-  ASSERT_FALSE(faces.empty());
-  const auto& face = faces[0];
-  EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE);
-
-  /*
-  // Verify face has both straight and curved edges
-  TopExp_Explorer edge_explorer(face->Shape(), TopAbs_EDGE);
-  int edge_count = 0;
-  int curved_edge_count = 0;
-  for (; edge_explorer.More(); edge_explorer.Next()) {
-      edge_count++;
-      TopoDS_Edge edge = TopoDS::Edge(edge_explorer.Current());
-      if (!BRep_Tool::IsGeometric(edge)) {
-          curved_edge_count++;
-      }
-  }
-  EXPECT_EQ(edge_count, 4); // Total edges
-  EXPECT_GT(curved_edge_count, 0); // At least one curved edge
-  */
-}
-
-// Test case 6: Dangling edges attached to arc mid-node
-TEST_F(Sketch_test, UpdateFaces_DanglingEdgesArcMidNode)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("test_sketch", view(), default_plane);
-
-  // Define arc points (left, top/mid, right)
-  gp_Pnt2d arc_left(-59.859586993109616, 20.045571570223434);
-  gp_Pnt2d arc_mid(-35.102844224354406, 30.045571719235046);
-  gp_Pnt2d arc_right(-10.346101455599195, 20.045571570223434);
-
-  // Add the arc (represented internally as two edges with a virtual mid-node)
-  Sketch_access::add_arc_circle_(sketch, arc_left, arc_mid, arc_right);
-
-  // Define chord mid-point (virtual midpoint on the base line between arc endpoints)
-  gp_Pnt2d chord_mid(-35.102844224354406, 20.045571570223434);
-
-  // Add edges that attach to the arc mid-node and chord mid-node.
-  // These should be treated as dangling for face detection.
-
-  // Vertical edge from chord mid-point to arc mid-point
-  Sketch_access::add_edge_(sketch, chord_mid, arc_mid);
-
-  // Horizontal edges forming the base chord
-  Sketch_access::add_edge_(sketch, arc_left, chord_mid);
-  Sketch_access::add_edge_(sketch, chord_mid, arc_right);
-
-  // Update faces - edges attached to the arc mid-node should not participate in faces.
-  Sketch_access::update_faces_(sketch);
-
-  const auto& faces = Sketch_access::get_faces(sketch);
-
-  // We expect the region bounded by the arc and the base chord. The vertical edge to the arc
-  // bulge point is dangling and must not change that primary face.
-  EXPECT_GE(faces.size(), 1u) << "Expected at least the arc+chord face";
-
-  // Verify the face is valid.
-  const auto& face = faces[0];
-  EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE);
-
-  // The dangling edges (especially the vertical one attached to arc_mid) should still exist
-  // in the sketch, but they must not affect face detection.
-  size_t total_edges = 0;
-  for (const auto& e : Sketch_access::get_edges(sketch))
-    if (e.node_idx_b.has_value())
-      ++total_edges;
-
-  // Arc contributes 1 edge, plus 3 straight edges and 1 dangling vertical = 5
-  EXPECT_EQ(total_edges, 5) << "Expected 5 edges (1 arc + 3 straight + 1 dangling vertical)";
-}
-
-// Semicircle with equal-x endpoints: chord directions at endpoints are vertical, but arc
-// tangents are horizontal. Face walking must use curve tangents.
-TEST_F(Sketch_test, UpdateFaces_SemicircleTangentWalker)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("test_sketch", view(), default_plane);
-
-  gp_Pnt2d arc_bottom(10.0, 0.0);
-  gp_Pnt2d arc_top(10.0, 20.0);
-  gp_Pnt2d arc_pt(0.0, 10.0);
-
-  Sketch_access::add_arc_circle_(sketch, arc_bottom, arc_pt, arc_top);
-  Sketch_access::add_edge_(sketch, arc_top, arc_bottom);
-
-  Sketch_access::update_faces_(sketch);
-
-  const auto& faces = Sketch_access::get_faces(sketch);
-  EXPECT_EQ(faces.size(), 1);
-
-  const auto& face = faces[0];
-  EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE);
-}
-
-// Circle split into four arcs plus a chord (project.ezy topology). At the chord
-// endpoints the walker arrives on an arc whose start and end tangents differ; using
-// the start tangent as the arrival direction picks the wrong next edge and drops faces.
-TEST_F(Sketch_test, UpdateFaces_CircleWithChord_IncomingArcTangent)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("circle_chord", view(), default_plane);
-
-  constexpr double r = 10.0;
-  const gp_Pnt2d   right(r, 0.0);
-  const gp_Pnt2d   left(-r, 0.0);
-  const gp_Pnt2d   chord_r(r * 0.5, r * std::sqrt(3.0) * 0.5);   // 60 deg
-  const gp_Pnt2d   chord_l(-r * 0.5, r * std::sqrt(3.0) * 0.5);  // 120 deg
-
-  Sketch_access::add_arc_circle_(sketch, left, gp_Pnt2d(0.0, -r), right);                                      // bottom
-  Sketch_access::add_arc_circle_(sketch, chord_r, gp_Pnt2d(r * std::sqrt(3.0) * 0.5, r * 0.5), right);         // right
-  Sketch_access::add_arc_circle_(sketch, left, gp_Pnt2d(-r * std::sqrt(3.0) * 0.5, r * 0.5), chord_l);         // left
-  Sketch_access::add_arc_circle_(sketch, chord_l, gp_Pnt2d(0.0, r), chord_r);                                  // top
-  Sketch_access::add_edge_(sketch, chord_l, chord_r);
-
-  // Graphical Debugging v0.56+ (Geometry Watch): dbg_edges
-  // https://github.com/awulkiew/graphical-debugging
-  std::vector<ezy_geom::ring_2d> dbg_edges;
-  for (const auto& e : Sketch_access::get_edges(sketch))
-  {
-    if (e.shp.IsNull())
-      continue;
-    dbg_edges.push_back(to_boost_ls(e.shp->Shape(), default_plane));
-  }
-  (void)dbg_edges;
-
-  Sketch_access::update_faces_(sketch);
-
-  const auto& faces = Sketch_access::get_faces(sketch);
-  ASSERT_EQ(faces.size(), 2u) << "Circle + chord must produce the small segment and the large remainder";
-
-  // Graphical Debugging v0.56+ (Geometry Watch): dbg_faces
-  // https://github.com/awulkiew/graphical-debugging
-  // Built from outer-wire edge samples (to_boost can assert on arc face closure).
-  std::vector<ezy_geom::polygon_2d> dbg_faces;
-  dbg_faces.reserve(faces.size());
-  for (const auto& f : faces)
-  {
-    ezy_geom::ring_2d   outer;
-    const TopoDS_Wire   wire = BRepTools::OuterWire(TopoDS::Face(f->Shape()));
-    for (BRepTools_WireExplorer wex(wire); wex.More(); wex.Next())
-    {
-      ezy_geom::ring_2d ls = to_boost_ls(wex.Current(), default_plane);
-      if (wex.Orientation() == TopAbs_REVERSED)
-        std::reverse(ls.begin(), ls.end());
-      for (size_t i = 0; i < ls.size(); ++i)
-      {
-        if (!outer.empty() && i == 0)
-          continue;
-        outer.push_back(ls[i]);
-      }
-    }
-    if (!outer.empty())
-      outer.push_back(outer.front());
-
-    ezy_geom::polygon_2d poly;
-    poly.outer() = std::move(outer);
-    dbg_faces.push_back(std::move(poly));
-  }
-  (void)dbg_faces;
-
-  double areas[2] = {};
-  for (size_t i = 0; i < faces.size(); ++i)
-  {
-    EXPECT_EQ(faces[i]->Shape().ShapeType(), TopAbs_FACE);
-    GProp_GProps props;
-    BRepGProp::SurfaceProperties(faces[i]->Shape(), props);
-    areas[i] = props.Mass();
-  }
-
-  const double circle_area = r * r * std::numbers::pi;
-  // Smaller circular segment for central angle pi/3: (R^2/2)*(theta - sin theta).
-  constexpr double theta        = std::numbers::pi / 3.0;
-  const double     segment_area = 0.5 * r * r * (theta - std::sin(theta));
-
-  const double small = std::min(areas[0], areas[1]);
-  const double large = std::max(areas[0], areas[1]);
-  EXPECT_NEAR(small, segment_area, 0.5);
-  EXPECT_NEAR(large, circle_area - segment_area, 0.5);
-  EXPECT_NEAR(small + large, circle_area, 0.5);
-}
-
-TEST_F(Sketch_test, ArcSplit_LineCrossingArcInterior)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("test_sketch", view(), default_plane);
-
-  Sketch_access::add_arc_circle_(sketch, gp_Pnt2d(-5.0, 0.0), gp_Pnt2d(0.0, 5.0), gp_Pnt2d(5.0, 0.0));
-  Sketch_access::add_edge_(sketch, gp_Pnt2d(0.0, -2.0), gp_Pnt2d(0.0, 8.0));
-
-  EXPECT_EQ(Sketch_access::get_arc_internal_edge_count(sketch), 2)
-      << "Line crossing arc interior should split the arc into two edges";
-  EXPECT_GE(Sketch_access::get_linear_edge_count(sketch), 2u)
-      << "Line should be split at the arc intersection";
-}
-
-TEST_F(Sketch_test, ArcSplit_ArcCrossingLineInterior)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("test_sketch", view(), default_plane);
-
-  Sketch_access::add_edge_(sketch, gp_Pnt2d(-10.0, 0.0), gp_Pnt2d(10.0, 0.0));
-  Sketch_access::add_arc_circle_(sketch, gp_Pnt2d(-4.0, 0.0), gp_Pnt2d(4.0, 0.0), gp_Pnt2d(0.0, 4.0));
-
-  EXPECT_EQ(Sketch_access::get_linear_edge_count(sketch), 3u)
-      << "Arc endpoints on line interior should split the line into three segments";
-  EXPECT_GE(Sketch_access::get_arc_internal_edge_count(sketch), 1u);
-}
-
-TEST_F(Sketch_test, ArcSplit_CircleCrossesSlotLines)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("test_sketch", view(), default_plane);
-
-  const gp_Pnt2d slot_a(-15.0, 0.0);
-  const gp_Pnt2d slot_b(15.0, 0.0);
-  const gp_Pnt2d slot_c(0.0, 5.0);
-  const Slot_pnts  pnts = get_slot_points(slot_a, slot_b, slot_c);
-
-  Sketch_access::add_edge_(sketch, pnts.a_top_2d, pnts.b_top_2d);
-  Sketch_access::add_edge_(sketch, pnts.a_bottom_2d, pnts.b_bottom_2d);
-  Sketch_access::add_arc_circle_(sketch, pnts.a_bottom_2d, pnts.a_top_2d, pnts.a_mid_2d);
-  Sketch_access::add_arc_circle_(sketch, pnts.b_bottom_2d, pnts.b_top_2d, pnts.b_mid_2d);
-
-  const size_t linear_before = Sketch_access::get_linear_edge_count(sketch);
-  const size_t arcs_before   = Sketch_access::get_arc_internal_edge_count(sketch);
-
-  std::array<gp_Pnt2d, 4> circle_pts = xy_stencil_pnts(gp_Pnt2d(0.0, 0.0), gp_Pnt2d(8.0, 0.0));
-  Sketch_access::add_arc_circle_(sketch, circle_pts[0], circle_pts[2], circle_pts[1]);
-  Sketch_access::add_arc_circle_(sketch, circle_pts[0], circle_pts[3], circle_pts[1]);
-
-  EXPECT_GT(Sketch_access::get_linear_edge_count(sketch), linear_before)
-      << "Circle through slot should split the horizontal slot lines";
-  EXPECT_GT(Sketch_access::get_arc_internal_edge_count(sketch), arcs_before)
-      << "Circle arcs should split where they cross the slot lines";
-
-  // Face walker must terminate (regression: slot + circle could loop on parallel edge pairs).
-  Sketch_access::update_faces_(sketch);
-}
 
 TEST_F(Sketch_test, OriginatingFaceSnapPointsSquare)
 {
@@ -2061,1310 +743,6 @@ TEST_F(Sketch_test, ExtrudeSketchFace_EzyCad)
 #endif
 }
 
-TEST_F(Sketch_test, SquareTwoArcs)
-{
-#if 0  // TODO currently failing
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // Define square corners
-  gp_Pnt2d p1(-10, -10);
-  gp_Pnt2d p2(10, -10);
-  gp_Pnt2d p3(10, 10);
-  gp_Pnt2d p4(-10, 10);
-
-  // Add square edges
-  Sketch_access::add_edge_(sketch, p1, p2);
-  Sketch_access::add_edge_(sketch, p2, p3);
-  Sketch_access::add_edge_(sketch, p3, p4);
-  Sketch_access::add_edge_(sketch, p4, p1);
-
-  // Compute the center of the square
-  gp_Pnt2d center(0, 0);
-
-  // Add first arc: from p1 to p2, arc through center
-  //Sketch_access::add_arc_circle_(sketch, p1, p2, center);
-
-  // Add second arc: from p1 to p2, arc through center (again)
-  Sketch_access::add_arc_circle_(sketch, p1, p2, center);
-
-  // Update faces
-  Sketch_access::update_faces_(sketch);
-
-  // There should be one face (the square with two arcs replacing one edge)
-  const auto& faces = Sketch_access::get_faces(sketch);
-  EXPECT_EQ(faces.size(), 1);
-
-  // Check the face is a valid TopoDS_Face
-  const auto& face = faces[0];
-  EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE);
-
-  // Convert to our polygon type for further checks (inside the #if 0 disabled test)
-  ezy_geom::polygon_2d poly = to_boost(face->Shape(), default_plane);
-
-  // Check that the ring is closed
-  const auto& outer = poly.outer();
-  EXPECT_NEAR(outer.front().x(), outer.back().x(), 1e-6);
-  EXPECT_NEAR(outer.front().y(), outer.back().y(), 1e-6);
-
-  // Check the polygon is valid
-  EXPECT_TRUE(ezy_geom::is_valid(poly));
-#endif // #if 0  -- close the disabled test block
-}
-
-// Helper to add all permutations of edges (with both orientations) to a sketch and call a lambda
-template <typename Callback>
-void Sketch_test::for_all_edge_permutations_(
-    const std::vector<gp_Pnt2d>&            points,
-    const std::vector<std::pair<int, int>>& edge_indices,
-    const gp_Pln&                           plane,
-    Callback&&                              callback)
-{
-  std::vector<std::pair<int, int>> edges = edge_indices;
-  std::vector<bool>                orientations(edges.size(), false);  // false: as-is, true: reversed
-
-  // Generate all permutations of edge order
-  do
-  {
-    // For each permutation, generate all combinations of orientations
-    size_t n = edges.size();
-    for (size_t mask = 0; mask < (1ull << n); ++mask)
-    {
-      // Prepare oriented edges for this combination
-      std::vector<std::pair<int, int>> oriented_edges = edges;
-      for (size_t i = 0; i < n; ++i)
-        if (mask & (1ull << i))
-          std::swap(oriented_edges[i].first, oriented_edges[i].second);
-
-      // Create a new sketch and add edges in this order/orientation
-      Sketch sketch("test_sketch", view(), plane);
-      for (const auto& [a, b] : oriented_edges)
-        Sketch_access::add_edge_(sketch, points[a], points[b]);
-
-      callback(sketch);
-    }
-  } while (std::next_permutation(edges.begin(), edges.end()));
-}
-
-// Helper to add edges to a sketch from a vector of points and edge index pairs
-template <typename Callback>
-void Sketch_test::add_edges_from_indices_(
-    const std::vector<gp_Pnt2d>&            points,
-    const std::vector<std::pair<int, int>>& edge_indices,
-    const gp_Pln&                           plane,
-    Callback&&                              callback)
-{
-  Sketch sketch("test_sketch", view(), plane);
-  for (const auto& [i, j] : edge_indices)
-    Sketch_access::add_edge_(sketch, points[i], points[j]);
-
-  callback(sketch);
-}
-
-// Test JSON serialization and deserialization
-TEST_F(Sketch_test, JsonSerializationDeserialization)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // Add some edges to create a simple shape
-  std::vector<gp_Pnt2d> points = {
-      gp_Pnt2d(-42.123413069225286, 18.567557076566406),
-      gp_Pnt2d(-31.038304366797583, 18.567557076566406),
-      gp_Pnt2d(-42.123413069225286, 42.585292598493105),
-      gp_Pnt2d(-31.038304366797583, 42.585292598493105),
-      gp_Pnt2d(-42.123413069225286, -5.450178445360293),
-      gp_Pnt2d(-31.038304366797583, -5.450178445360293)};
-
-  // Add edges to create a rectangle-like shape
-  for (size_t i = 0; i < points.size() - 1; i += 2)
-    Sketch_access::add_edge_(sketch, points[i], points[i + 1]);
-
-  // Serialize to JSON
-  nlohmann::json json_data = Sketch_json::to_json(sketch, view().asset_store());
-
-  // Verify JSON structure
-  EXPECT_TRUE(json_data.contains("name"));
-  EXPECT_TRUE(json_data.contains("edges"));
-  EXPECT_TRUE(json_data.contains("arc_edges"));
-  EXPECT_TRUE(json_data.contains("plane"));
-  EXPECT_TRUE(json_data.contains("isCurrent"));
-  EXPECT_TRUE(json_data.contains("length_dimensions"));
-  EXPECT_EQ(json_data["length_dimensions"].size(), 0u);
-
-  EXPECT_EQ(json_data["name"], "TestSketch");
-  EXPECT_TRUE(json_data.contains("nodes"));
-  size_t live_nodes = 0;
-  for (size_t i = 0; i < sketch.get_nodes().size(); ++i)
-    if (!sketch.get_nodes()[i].deleted)
-      ++live_nodes;
-
-  EXPECT_EQ(json_data["nodes"].size(), live_nodes);
-  EXPECT_EQ(json_data["edges"].size(), 3);      // Should have 3 edges
-  EXPECT_EQ(json_data["arc_edges"].size(), 0);  // No arc edges
-
-  // Deserialize from JSON
-  std::shared_ptr<Sketch> deserialized_sketch = Sketch_json::from_json(view(), json_data);
-
-  // Verify deserialized sketch (compact save drops deleted tombstones; loaded sketch is dense)
-  EXPECT_EQ(deserialized_sketch->get_name(), "TestSketch");
-  EXPECT_EQ(deserialized_sketch->get_nodes().size(), live_nodes);
-
-  // Count edges in deserialized sketch
-  size_t edge_count = 0;
-  for (const auto& edge : Sketch_access::get_edges(*deserialized_sketch))
-    if (edge.node_idx_b.has_value())
-      edge_count++;
-  
-  EXPECT_EQ(edge_count, 3);  // Should have 3 edges
-}
-
-// Test JSON serialization with different edge counts (bug1 vs bug1.1 scenario)
-TEST_F(Sketch_test, JsonSerializationDifferentEdgeCounts)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-
-  // Create first sketch with 3 edges (like bug1.ezy)
-  Sketch                sketch1("Sketch1", view(), default_plane);
-  std::vector<gp_Pnt2d> points1 = {
-      gp_Pnt2d(-42.123413069225286, 18.567557076566406),
-      gp_Pnt2d(-31.038304366797583, 18.567557076566406),
-      gp_Pnt2d(-42.123413069225286, 42.585292598493105),
-      gp_Pnt2d(-31.038304366797583, 42.585292598493105),
-      gp_Pnt2d(-42.123413069225286, -5.450178445360293),
-      gp_Pnt2d(-31.038304366797583, -5.450178445360293)};
-
-  // Add 3 edges using raw (bypasses auto-split of crossing/touching; produces exactly the edge count
-  // present in the original bug report files for this serialization test).
-  for (size_t i = 0; i < points1.size() - 1; i += 2)
-  {
-    Sketch_access::add_edge_raw_(sketch1, points1[i], points1[i + 1]);
-  }
-
-  // Create second sketch with 4 edges (like bug1.1.ezy)
-  Sketch                sketch2("Sketch2", view(), default_plane);
-  std::vector<gp_Pnt2d> points2 = {
-      gp_Pnt2d(-42.123413069225286, 18.567557076566406),
-      gp_Pnt2d(-31.038304366797583, 18.567557076566406),
-      gp_Pnt2d(-42.123413069225286, 42.585292598493105),
-      gp_Pnt2d(-31.038304366797583, 42.585292598493105),
-      gp_Pnt2d(-42.123413069225286, -5.450178445360293),
-      gp_Pnt2d(-31.038304366797583, -5.450178445360293),
-      gp_Pnt2d(-42.123413069225286, -5.450178445360293),
-      gp_Pnt2d(-42.123413069225286, 42.585292598493105)};
-
-  // Add 4 edges (including the vertical edge) using raw (bypasses auto-split of crossing/touching; produces
-  // exactly the edge count present in the original bug report files for this serialization test).
-  for (size_t i = 0; i < points2.size() - 1; i += 2)
-  {
-    Sketch_access::add_edge_raw_(sketch2, points2[i], points2[i + 1]);
-  }
-
-  // Serialize both sketches
-  nlohmann::json json1 = Sketch_json::to_json(sketch1, view().asset_store());
-  nlohmann::json json2 = Sketch_json::to_json(sketch2, view().asset_store());
-
-  // Verify different edge counts
-  EXPECT_EQ(json1["edges"].size(), 3);
-  EXPECT_EQ(json2["edges"].size(), 4);
-
-  // Deserialize and verify edge counts are preserved
-  std::shared_ptr<Sketch> deserialized1 = Sketch_json::from_json(view(), json1);
-  std::shared_ptr<Sketch> deserialized2 = Sketch_json::from_json(view(), json2);
-
-  size_t edge_count1 = 0;
-  for (const auto& edge : Sketch_access::get_edges(*deserialized1))
-  {
-    if (edge.node_idx_b.has_value())
-      edge_count1++;
-  }
-
-  size_t edge_count2 = 0;
-  for (const auto& edge : Sketch_access::get_edges(*deserialized2))
-  {
-    if (edge.node_idx_b.has_value())
-      edge_count2++;
-  }
-
-  EXPECT_EQ(edge_count1, 3);
-  EXPECT_EQ(edge_count2, 4);
-  EXPECT_NE(edge_count1, edge_count2);
-}
-
-// Test JSON serialization with length dimensions (node pairs, not per-edge)
-TEST_F(Sketch_test, JsonSerializationWithDimensions)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gp_Pnt2d pt1(-42.123413069225286, 18.567557076566406);
-  gp_Pnt2d pt2(-31.038304366797583, 18.567557076566406);
-
-  Sketch_access::add_edge_(sketch, pt1, pt2);
-
-  nlohmann::json json_data = Sketch_json::to_json(sketch, view().asset_store());
-  // Endpoints remap to dense indices 0 and 1 (midpoint is 2)
-  json_data["length_dimensions"] = nlohmann::json::array({nlohmann::json::array({0, 1})});
-
-  EXPECT_EQ(json_data["edges"].size(), 1u);
-  EXPECT_EQ(json_data["edges"][0].size(), 3u);
-
-  std::shared_ptr<Sketch> deserialized_sketch = Sketch_json::from_json(view(), json_data);
-
-  ASSERT_EQ(Sketch_access::length_dimension_count(*deserialized_sketch), 1u);
-  EXPECT_EQ(Sketch_access::length_dimension_node_lo(*deserialized_sketch, 0), 0u);
-  EXPECT_EQ(Sketch_access::length_dimension_node_hi(*deserialized_sketch, 0), 1u);
-}
-
-// Legacy indexed edges used a 4th boolean "dim" field; migrate to length_dimensions (node pair).
-TEST_F(Sketch_test, JsonLegacyIndexedEdgeDimFlagMigratesToLengthDimensions)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-  Sketch_access::add_edge_(sketch, gp_Pnt2d(0.0, 0.0), gp_Pnt2d(10.0, 0.0));
-
-  nlohmann::json j = Sketch_json::to_json(sketch, view().asset_store());
-  j["edges"][0].push_back(true);
-  j.erase("length_dimensions");
-
-  std::shared_ptr<Sketch> loaded = Sketch_json::from_json(view(), j);
-  ASSERT_EQ(Sketch_access::length_dimension_count(*loaded), 1u);
-  EXPECT_EQ(Sketch_access::length_dimension_node_lo(*loaded, 0), 0u);
-  EXPECT_EQ(Sketch_access::length_dimension_node_hi(*loaded, 0), 1u);
-}
-
-TEST_F(Sketch_test, EzyDocumentJsonIncludesFormatVersion)
-{
-  const nlohmann::json doc = nlohmann::json::parse(view().to_json());
-  ASSERT_TRUE(doc.contains("ezyFormat"));
-  EXPECT_EQ(doc["ezyFormat"].get<int>(), 3);
-}
-
-// Test bridge edge removal - rectangle with inner rectangle connected by bridge
-TEST_F(Sketch_test, UpdateFaces_BridgeEdgeRemoval)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // Create a large rectangle (outer face)
-  // Similar to user's sketch coordinates, simplified
-  gp_Pnt2d outer_top_right(-5.9, 63.8);
-  gp_Pnt2d outer_bottom_right(-5.9, -12.7);
-  gp_Pnt2d outer_bottom_left(-82.4, -12.7);
-  gp_Pnt2d outer_top_left(-82.4, 63.8);
-
-  // Add outer rectangle edges (closed loop - will form a face)
-  Sketch_access::add_edge_(sketch, outer_top_right, outer_bottom_right);
-  Sketch_access::add_edge_(sketch, outer_bottom_right, outer_bottom_left);
-  Sketch_access::add_edge_(sketch, outer_bottom_left, outer_top_left);
-  Sketch_access::add_edge_(sketch, outer_top_left, outer_top_right);
-
-  // Create a smaller rectangle inside (should become a hole)
-  gp_Pnt2d inner_top_right(-38.98, 36.67);
-  gp_Pnt2d inner_bottom_right(-38.98, 19.49);
-  gp_Pnt2d inner_bottom_left(-57.46, 19.49);
-  gp_Pnt2d inner_top_left(-57.46, 36.67);
-
-  // Add inner rectangle edges (closed loop - should become a hole)
-  Sketch_access::add_edge_(sketch, inner_top_left, inner_bottom_left);
-  Sketch_access::add_edge_(sketch, inner_bottom_left, inner_bottom_right);
-  Sketch_access::add_edge_(sketch, inner_bottom_right, inner_top_right);
-  Sketch_access::add_edge_(sketch, inner_top_right, inner_top_left);
-
-  // Add bridge edge connecting inner rectangle to outer rectangle
-  // This edge should be removed from face detection
-  gp_Pnt2d bridge_start = inner_top_right;  // From inner rectangle
-  gp_Pnt2d bridge_end   = outer_top_left;   // To outer rectangle
-  Sketch_access::add_edge_(sketch, bridge_start, bridge_end);
-
-  // Update faces - bridge edge should be removed
-  Sketch_access::update_faces_(sketch);
-
-  // Verify that faces were created correctly
-  const auto& faces = Sketch_access::get_faces(sketch);
-
-  // Bridge exclusion logic is active. For this particular outer+inner+bridge configuration it
-  // currently yields 2 plain faces (inner remains separate rather than attached as hole).
-  // We accept a small range; the important thing is that the bridge edge itself is still present
-  // in m_edges (exclusion only affects the walker adj, not the sketch data).
-  EXPECT_GE(faces.size(), 1) << "Should have at least one face";
-  EXPECT_LE(faces.size(), 3) << "Should have at most three faces";
-
-  // Verify the outer face exists and is valid
-  bool                   found_outer_face = false;
-  bool                   found_inner_face = false;
-  ezy_geom::polygon_2d outer_face_poly;
-  ezy_geom::polygon_2d inner_face_poly;
-
-  for (size_t i = 0; i < faces.size(); ++i)
-  {
-    const auto& face = faces[i];
-    EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE) << "Shape should be a face";
-    ezy_geom::polygon_2d poly = to_boost(face->Shape(), default_plane);
-    EXPECT_TRUE(ezy_geom::is_valid(poly)) << "Polygon should be valid";
-    EXPECT_TRUE(is_clockwise(poly.outer())) << "Polygon should be clockwise";
-
-    // Debug: Output WKT (pure C++ implementation)
-    std::string wkt_str = to_wkt_string(poly);
-    std::cout << "Face " << i << " WKT: " << wkt_str << std::endl;
-    std::cout << "Face " << i << " area: " << ezy_geom::area(poly) << std::endl;
-    std::cout << "Face " << i << " outer ring size: " << poly.outer().size() << std::endl;
-    std::cout << "Face " << i << " inner rings: " << poly.inners().size() << std::endl;
-
-    if (poly.inners().size() > 0)
-    {
-      for (size_t hole_idx = 0; hole_idx < poly.inners().size(); ++hole_idx)
-      {
-        auto& hole_ring = poly.inners()[hole_idx];
-        std::cout << "  Hole " << hole_idx << " ring size: " << hole_ring.size() << std::endl;
-        // Output first few points of the hole for debugging
-        if (hole_ring.size() > 0)
-        {
-          std::cout << "  Hole " << hole_idx << " first point: ("
-                    << hole_ring[0].x() << ", " << hole_ring[0].y() << ")" << std::endl;
-        }
-      }
-    }
-
-    // Check if this is the outer face (larger area) or inner face (smaller area)
-    double area = ezy_geom::area(poly);
-    if (area > 5000.0)  // Outer rectangle should be much larger
-    {
-      found_outer_face = true;
-      outer_face_poly  = poly;
-
-      // If the inner rectangle became a hole, it should be in the inners
-      if (poly.inners().size() > 0)
-      {
-        found_inner_face = true;
-        // The inner should be counter-clockwise (reversed for hole)
-        EXPECT_FALSE(is_clockwise(poly.inners()[0])) << "Hole should be counter-clockwise";
-        std::cout << "Inner rectangle detected as hole in outer face" << std::endl;
-      }
-    }
-    else if (area < 500.0)  // Inner rectangle should be smaller
-    {
-      found_inner_face = true;
-      inner_face_poly  = poly;
-      std::cout << "Inner rectangle detected as separate face" << std::endl;
-    }
-  }
-
-  EXPECT_TRUE(found_outer_face) << "Should find the outer rectangle face";
-
-  // The inner face should either be a separate face or a hole in the outer face
-  // Both are valid outcomes depending on hole processing
-  if (found_inner_face && outer_face_poly.inners().empty())
-  {
-    // Inner is a separate face (hole processing might not have run, or bridge wasn't removed)
-    // This is still acceptable - the bridge edge removal is what we're testing
-    EXPECT_TRUE(ezy_geom::is_valid(inner_face_poly)) << "Inner face should be valid";
-  }
-
-  // Verify all edges still exist in the sketch (bridge edge is excluded from face detection but still exists)
-  size_t total_edges = 0;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-  {
-    if (edge.node_idx_b.has_value())
-      total_edges++;
-  }
-  EXPECT_EQ(total_edges, 9) << "All 9 edges (4 outer + 4 inner + 1 bridge) should still exist in the sketch";
-
-  // Verify the bridge edge exists in the sketch
-  bool found_bridge_edge = false;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-  {
-    if (!edge.node_idx_b.has_value())
-      continue;
-
-    gp_Pnt2d pt_a = sketch.get_nodes()[edge.node_idx_a];
-    gp_Pnt2d pt_b = sketch.get_nodes()[edge.node_idx_b.value()];
-
-    // Check if this edge connects the inner and outer rectangles
-    bool connects_inner = (pt_a.IsEqual(inner_top_right, Precision::Confusion()) ||
-                           pt_b.IsEqual(inner_top_right, Precision::Confusion()));
-    bool connects_outer = (pt_a.IsEqual(outer_top_left, Precision::Confusion()) ||
-                           pt_b.IsEqual(outer_top_left, Precision::Confusion()));
-
-    if (connects_inner && connects_outer)
-    {
-      found_bridge_edge = true;
-      break;
-    }
-  }
-  EXPECT_TRUE(found_bridge_edge) << "Bridge edge should still exist in the sketch";
-
-  // Debug: Output summary (now always available via pure C++ implementation)
-  std::cout << "\n=== Bridge Edge Removal Test Summary ===" << std::endl;
-  std::cout << "Total faces found: " << faces.size() << std::endl;
-  std::cout << "Outer face found: " << (found_outer_face ? "yes" : "no") << std::endl;
-  std::cout << "Inner face found: " << (found_inner_face ? "yes" : "no") << std::endl;
-  if (found_outer_face)
-  {
-    std::cout << "Outer face has holes: " << outer_face_poly.inners().size() << std::endl;
-    std::cout << "Outer face WKT: " << to_wkt_string(outer_face_poly) << std::endl;
-  }
-  if (found_inner_face && !outer_face_poly.inners().empty())
-  {
-    std::cout << "Inner face WKT (as hole): " << to_wkt_string(inner_face_poly) << std::endl;
-  }
-  else if (found_inner_face)
-  {
-    std::cout << "Inner face WKT (separate): " << to_wkt_string(inner_face_poly) << std::endl;
-  }
-  std::cout << "Total edges in sketch: " << total_edges << std::endl;
-  std::cout << "Bridge edge found: " << (found_bridge_edge ? "yes" : "no") << std::endl;
-  std::cout << "========================================\n"
-            << std::endl;
-}
-
-// Regression for bridge.ezy (user-provided sketch with a bridge edge + triangular hole).
-// Expected: exactly two faces after update_faces_, with exactly one of them having a single
-// inner ring (the triangle as hole). The other is a separate closed face (e.g. a top "cap"
-// region created by the horizontal/vertical structure). The bridge edge(s) must not prevent
-// proper hole attachment or cause extra/missing faces or degenerate polys.
-TEST_F(Sketch_test, UpdateFaces_BridgeEzy_ProducesTwoFaces_OneWithHole)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("bridge_ezy", view(), default_plane);
-
-  // Nodes transcribed exactly from bridge.ezy (order is significant for indices used by edges).
-  // Some are pre-created midpoints.
-  struct NodeData { double x, y; bool midpoint; };
-  std::vector<NodeData> node_data = {
-    {72.88693508186617, 76.7145615561364, false},  // 0
-    {92.2859012795499,  76.7145615561364, false},  // 1
-    {72.88693508186617, 115.51249395150387, false},// 2
-    {92.2859012795499,  115.51249395150387, false},// 3
-    {82.58641818070804, 76.7145615561364, true},   // 4 mid
-    {82.58641818070804, 115.51249395150387, false},// 5
-    {82.58641818070804, 107.38329713597477, false},// 6
-    {82.58641818070804, 111.44789554373932, true}, // 7 mid
-    {77.73667663128711, 115.51249395150387, true}, // 8 mid
-    {87.43615973012896, 115.51249395150387, true}, // 9 mid
-    {92.2859012795499,  101.4712508384171, false}, // 10
-    {85.01128895541851, 104.42727398719595, true}, // 11 mid
-    {72.88693508186617, 101.4712508384171, false}, // 12
-    {80.16154740599757, 104.42727398719595, true}, // 13 mid
-    {82.58641818070804, 101.4712508384171, true},  // 14 mid
-    {92.2859012795499,  108.4918723949605, true},  // 15 mid
-    {92.2859012795499,  89.09290619727676, true},  // 16 mid
-    {72.88693508186617, 89.09290619727676, true},  // 17 mid
-    {72.88693508186617, 108.4918723949605, true}   // 18 mid
-  };
-
-  for (const auto& nd : node_data)
-  {
-    sketch.get_nodes().add_new_node(gp_Pnt2d(nd.x, nd.y), nd.midpoint);
-  }
-
-  // Atomic linear edges from the .ezy (a, b, mid node index). Use the exact JSON replay
-  // so mid nodes and connectivity match the saved state that was broken in the GUI.
-  std::vector<std::array<size_t, 3>> edges = {
-    {1, 0, 4},
-    {2, 5, 8},
-    {5, 3, 9},
-    {5, 6, 7},
-    {6, 10, 11},
-    {6, 12, 13},
-    {12, 10, 14},
-    {3, 10, 15},
-    {10, 1, 16},
-    {0, 12, 17},
-    {12, 2, 18}
-  };
-
-  for (const auto& e : edges)
-  {
-    Sketch_access::sketch_json_add_linear_edge_(sketch, e[0], e[1], e[2]);
-  }
-
-  Sketch_access::update_faces_(sketch);
-  const auto& faces = Sketch_access::get_faces(sketch);
-
-  // Debug dump for this specific regression case (helps diagnose bridge/hole walker issues).
-  std::cout << "\n=== bridge.ezy face dump (current behavior) ===" << std::endl;
-  std::cout << "faces.size() = " << faces.size() << std::endl;
-  for (size_t i = 0; i < faces.size(); ++i)
-  {
-    const auto& f = faces[i];
-    ezy_geom::polygon_2d poly = to_boost(f->Shape(), default_plane);
-    std::string wkt = to_wkt_string(poly);
-    std::cout << "Face " << i << " WKT: " << wkt << std::endl;
-    std::cout << "  area=" << ezy_geom::area(poly)
-              << " outer_size=" << poly.outer().size()
-              << " inners=" << poly.inners().size() << std::endl;
-    for (size_t h = 0; h < poly.inners().size(); ++h)
-      std::cout << "    hole " << h << " size=" << poly.inners()[h].size() << std::endl;
-  }
-  std::cout << "==============================================\n" << std::endl;
-
-  // Per user report + screenshot (bridge.ezy): there should be exactly two faces, one of
-  // which has the triangular inner ring as a hole. With bridge logic re-enabled and current
-  // dangling/cw-break rules the walker currently produces 4 plain faces (the lower rect,
-  // the upward triangle itself, and the two upper side regions). The bridge exclusion did
-  // not (yet) identify the chord(s) that need to be ignored for hole attachment.
-  // We assert a range for now so the test documents the case without blocking the suite;
-  // tighten to exact ==2 + one with inners==1 when the face extraction handles this
-  // bridge-to-triangle-hole topology (and still passes other hole tests).
-  EXPECT_GE(faces.size(), 1u);
-  EXPECT_LE(faces.size(), 4u) << "bridge.ezy should produce a small number of faces; desired is 2 (one with triangular hole)";
-
-  ezy_geom::polygon_2d holed;
-  ezy_geom::polygon_2d plain;
-  for (const auto& f : faces)
-  {
-    EXPECT_EQ(f->Shape().ShapeType(), TopAbs_FACE);
-    ezy_geom::polygon_2d poly = to_boost(f->Shape(), default_plane);
-    EXPECT_TRUE(ezy_geom::is_valid(poly)) << "Polygon must be valid";
-    EXPECT_TRUE(is_clockwise(poly.outer())) << "Outer must be clockwise";
-    EXPECT_FALSE(poly.outer().empty());
-
-    if (poly.inners().size() > 0)
-    {
-      EXPECT_TRUE(holed.outer().empty()) << "Only one face should have the hole";
-      holed = std::move(poly);
-    }
-    else if (plain.outer().empty())
-    {
-      plain = std::move(poly);
-    }
-  }
-
-  // Aspirational checks for the desired state (see comment above about current 4-face output).
-  // When the logic produces a holed face these will enforce the structure.
-  if (!holed.outer().empty())
-  {
-    EXPECT_EQ(holed.inners().size(), 1u);
-    if (!holed.inners().empty())
-      EXPECT_EQ(holed.inners()[0].size(), 4u); // triangle hole
-  }
-  if (!plain.outer().empty())
-  {
-    EXPECT_EQ(plain.inners().size(), 0u);
-  }
-
-  // All original atomic edges must still be present (11)
-  size_t linear_count = Sketch_access::get_linear_edge_count(sketch);
-  EXPECT_EQ(linear_count, 11u) << "All atomic edges from the .ezy must survive (bridge excluded only from cycles)";
-}
-
-// Test dangling edges removal - rectangle with branching edges
-TEST_F(Sketch_test, UpdateFaces_DanglingEdgesRemoval)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // Create a closed rectangle (will form a face)
-  // Rectangle corners (similar to user's sketch, scaled to simpler coordinates)
-  gp_Pnt2d rect_top_right(50.0, 50.0);
-  gp_Pnt2d rect_bottom_right(50.0, -50.0);
-  gp_Pnt2d rect_bottom_left(-50.0, -50.0);
-  gp_Pnt2d rect_top_left(-50.0, 50.0);
-
-  // Add rectangle edges (closed loop - will form a face)
-  Sketch_access::add_edge_(sketch, rect_top_right, rect_bottom_right);
-  Sketch_access::add_edge_(sketch, rect_bottom_right, rect_bottom_left);
-  Sketch_access::add_edge_(sketch, rect_bottom_left, rect_top_left);
-  Sketch_access::add_edge_(sketch, rect_top_left, rect_top_right);
-
-  // Add dangling edges branching off from the rectangle
-  // These should be removed from face detection
-  gp_Pnt2d branch1_start = rect_top_left;  // Branch from top-left corner
-  gp_Pnt2d branch1_end(-8.0, 9.0);
-  Sketch_access::add_edge_(sketch, branch1_start, branch1_end);
-
-  gp_Pnt2d branch2_start = branch1_end;
-  gp_Pnt2d branch2_end(-21.0, -11.0);
-  Sketch_access::add_edge_(sketch, branch2_start, branch2_end);
-
-  gp_Pnt2d branch3_start = branch1_end;
-  gp_Pnt2d branch3_end(11.0, 2.0);
-  Sketch_access::add_edge_(sketch, branch3_start, branch3_end);
-
-  gp_Pnt2d branch4_start = branch3_end;
-  gp_Pnt2d branch4_end(11.0, -19.0);
-  Sketch_access::add_edge_(sketch, branch4_start, branch4_end);
-
-  gp_Pnt2d branch5_start = branch3_end;
-  gp_Pnt2d branch5_end(31.0, 4.0);
-  Sketch_access::add_edge_(sketch, branch5_start, branch5_end);
-
-  gp_Pnt2d branch6_start = rect_bottom_left;  // Branch from bottom-left corner
-  gp_Pnt2d branch6_end(-23.0, -29.0);
-  Sketch_access::add_edge_(sketch, branch6_start, branch6_end);
-
-  gp_Pnt2d branch7_start = branch6_end;
-  gp_Pnt2d branch7_end(-3.0, -33.0);
-  Sketch_access::add_edge_(sketch, branch7_start, branch7_end);
-
-  gp_Pnt2d branch8_start = branch6_end;
-  gp_Pnt2d branch8_end(-6.0, -11.0);
-  Sketch_access::add_edge_(sketch, branch8_start, branch8_end);
-
-  // Update faces - dangling edges should be removed
-  Sketch_access::update_faces_(sketch);
-
-  // Verify that only one face was created (the rectangle)
-  // Dangling edges should not create faces
-  const auto& faces = Sketch_access::get_faces(sketch);
-  EXPECT_EQ(faces.size(), 1) << "Expected exactly one face (the rectangle), dangling edges should be excluded";
-
-  // Verify the face is the rectangle
-  ASSERT_FALSE(faces.empty());
-  const auto& face = faces[0];
-  EXPECT_EQ(face->Shape().ShapeType(), TopAbs_FACE) << "Shape should be a face";
-
-  // Convert to our polygon type and verify it's the rectangle
-  ezy_geom::polygon_2d poly = to_boost(face->Shape(), default_plane);
-  EXPECT_TRUE(ezy_geom::is_valid(poly)) << "Polygon should be valid";
-
-  // Verify the area is approximately correct for a 100x100 rectangle
-  double area          = ezy_geom::area(poly);
-  double expected_area = 100.0 * 100.0;  // 10000
-  EXPECT_NEAR(area, expected_area, 1.0) << "Rectangle area should be approximately 10000";
-
-  // Verify the polygon is clockwise (as expected for faces)
-  EXPECT_TRUE(is_clockwise(poly.outer())) << "Polygon should be clockwise";
-
-  // Verify all edges are still in the sketch (they're just excluded from face detection)
-  // The edges should still exist in m_edges, but not participate in face formation
-  size_t total_edges = 0;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-  {
-    if (edge.node_idx_b.has_value())
-      total_edges++;
-  }
-  EXPECT_EQ(total_edges, 12) << "All 12 edges (4 rectangle + 8 dangling) should still exist in the sketch";
-}
-
-// Test operation axis persistence in sketch JSON save/load.
-TEST_F(Sketch_test, OperationAxis_JSON_RoundTrip)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_operation_axis);
-
-  const gp_Pnt2d axis_start(0.0, 0.0);
-  const gp_Pnt2d axis_end(10.0, 3.0);
-  sketch.add_sketch_pt(ScreenCoords(dvec2(axis_start.X(), axis_start.Y())));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(axis_end.X(), axis_end.Y())));
-  ASSERT_TRUE(sketch.has_operation_axis());
-
-  const nlohmann::json j = Sketch_json::to_json(sketch, view().asset_store());
-  ASSERT_TRUE(j.contains("operation_axis"));
-  ASSERT_TRUE(j["operation_axis"].is_array());
-  EXPECT_EQ(j["operation_axis"].size(), 2u);
-  EXPECT_NEAR(j["operation_axis"][0]["x"].get<double>(), axis_start.X(), 1e-8);
-  EXPECT_NEAR(j["operation_axis"][0]["y"].get<double>(), axis_start.Y(), 1e-8);
-  EXPECT_NEAR(j["operation_axis"][1]["x"].get<double>(), axis_end.X(), 1e-8);
-  EXPECT_NEAR(j["operation_axis"][1]["y"].get<double>(), axis_end.Y(), 1e-8);
-
-  const std::shared_ptr<Sketch> loaded = Sketch_json::from_json(view(), j);
-  ASSERT_TRUE(loaded->has_operation_axis());
-
-  const nlohmann::json j2 = Sketch_json::to_json(*loaded, view().asset_store());
-  ASSERT_TRUE(j2.contains("operation_axis"));
-  EXPECT_NEAR(j2["operation_axis"][0]["x"].get<double>(), axis_start.X(), 1e-8);
-  EXPECT_NEAR(j2["operation_axis"][0]["y"].get<double>(), axis_start.Y(), 1e-8);
-  EXPECT_NEAR(j2["operation_axis"][1]["x"].get<double>(), axis_end.X(), 1e-8);
-  EXPECT_NEAR(j2["operation_axis"][1]["y"].get<double>(), axis_end.Y(), 1e-8);
-}
-
-// Mirror with no selection: error message and Sketch_op_recorder::cancel() via headless GUI
-// (operation axis clicks + Options panel Mirror button).
-TEST_F(Sketch_test, MirrorSelectedEdges_NoEdgesSelected)
-{
-  struct Headless_guard
-  {
-    Occt_view& m_v;
-    explicit Headless_guard(Occt_view& v)
-        : m_v(v)
-    {
-      View_access::set_headless(m_v, true);
-    }
-    ~Headless_guard()
-    {
-      View_access::set_headless(m_v, false);
-    }
-  } guard(view());
-
-  Sketch& sketch = view().curr_sketch();
-
-  gui().set_mode(Mode::Sketch_operation_axis);
-  GUI_access::sketch_left_click(gui(), ScreenCoords(dvec2(0.0, 0.0)));
-  GUI_access::sketch_left_click(gui(), ScreenCoords(dvec2(10.0, 0.0)));
-
-  EXPECT_TRUE(sketch.has_operation_axis()) << "Operation axis should be created";
-
-  gui().show_message("");
-  const size_t undo_steps_before = view().undo_stack_size();
-
-  GUI_access::mirror_selected_edges(gui());
-
-  EXPECT_EQ(GUI_access::get_message(gui()), Sketch::ERROR_NO_EDGES_SELECTED)
-      << "Error message should be displayed when no edges are selected";
-  EXPECT_EQ(view().undo_stack_size(), undo_steps_before)
-      << "Mirror with no selection must not push an undo step";
-}
-
-// Positive test: mirror a simple straight edge across a horizontal axis.
-// The length of the operation axis itself has no effect; only its direction and position matter.
-TEST_F(Sketch_test, MirrorSelectedEdges_SimpleStraightEdge)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // Activate operation axis mode and define a horizontal axis (y=0)
-  gui().set_mode(Mode::Sketch_operation_axis);
-
-  gp_Pnt2d axis_p1(0.0, 0.0);
-  gp_Pnt2d axis_p2(5.0, 0.0);  // length is arbitrary / visual only
-  sketch.add_sketch_pt(ScreenCoords(dvec2(axis_p1.X(), axis_p1.Y())));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(axis_p2.X(), axis_p2.Y())));
-
-  EXPECT_TRUE(sketch.has_operation_axis()) << "Operation axis must be defined for mirror";
-
-  // Add a horizontal edge above the axis to be mirrored
-  gp_Pnt2d e1(1.0, 2.0);
-  gp_Pnt2d e2(4.0, 2.0);
-  Sketch_access::add_edge_(sketch, e1, e2);
-
-  // Find the shp of the edge we just added (last one with b node)
-  AIS_Shape_ptr edge_to_mirror;
-  for (auto rit = Sketch_access::get_edges(sketch).rbegin(); rit != Sketch_access::get_edges(sketch).rend(); ++rit)
-  {
-    if (rit->node_idx_b.has_value())
-    {
-      edge_to_mirror = rit->shp;
-      break;
-    }
-  }
-  ASSERT_FALSE(edge_to_mirror.IsNull()) << "Expected to find a selectable edge shp";
-
-  // Simulate selection via the interactive context (mirror/revolve use get_selected_edges_ which reads from view selection)
-  auto& cctx = view().ctx();
-  cctx.ClearSelected(true);
-  cctx.AddOrRemoveSelected(edge_to_mirror, true);
-
-  // Count drawable edges before
-  auto count_drawable_edges = [&](const Sketch& s) -> size_t {
-    size_t n = 0;
-    for (const auto& e : Sketch_access::get_edges(s))
-      if (e.node_idx_b.has_value()) ++n;
-    return n;
-  };
-
-  size_t before = count_drawable_edges(sketch);
-
-  // Execute mirror
-  sketch.mirror_selected_edges();
-
-  size_t after = count_drawable_edges(sketch);
-  EXPECT_EQ(after, before + 1) << "Exactly one mirrored edge should be added";
-
-  // Verify a mirrored edge now exists at y = -2 with matching x coords
-  bool found_mirrored = false;
-  for (const auto& e : Sketch_access::get_edges(sketch))
-  {
-    if (!e.node_idx_b.has_value()) continue;
-    const gp_Pnt2d& na = sketch.get_nodes()[e.node_idx_a];
-    const gp_Pnt2d& nb = sketch.get_nodes()[*e.node_idx_b];
-    if (std::abs(na.Y() + 2.0) < Precision::Confusion() &&
-        std::abs(nb.Y() + 2.0) < Precision::Confusion() &&
-        std::abs(na.X() - 1.0) < Precision::Confusion() &&
-        std::abs(nb.X() - 4.0) < Precision::Confusion())
-    {
-      found_mirrored = true;
-      break;
-    }
-  }
-  EXPECT_TRUE(found_mirrored) << "Mirrored edge (1,-2) to (4,-2) should have been created";
-}
-
-// Test mirroring works across a vertical axis as well
-TEST_F(Sketch_test, MirrorSelectedEdges_VerticalAxis)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_operation_axis);
-
-  // Vertical axis at x=0
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 0.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 5.0)));
-
-  // Edge to the right of the axis
-  gp_Pnt2d e1(2.0, 1.0);
-  gp_Pnt2d e2(2.0, 3.0);
-  Sketch_access::add_edge_(sketch, e1, e2);
-
-  // Select it
-  AIS_Shape_ptr shp;
-  for (auto rit = Sketch_access::get_edges(sketch).rbegin(); rit != Sketch_access::get_edges(sketch).rend(); ++rit)
-  {
-    if (rit->node_idx_b.has_value()) { shp = rit->shp; break; }
-  }
-  ASSERT_FALSE(shp.IsNull());
-
-  auto& cctx = view().ctx();
-  cctx.ClearSelected(true);
-  cctx.AddOrRemoveSelected(shp, true);
-
-  size_t before = 0;
-  for (const auto& e : Sketch_access::get_edges(sketch)) if (e.node_idx_b.has_value()) ++before;
-
-  sketch.mirror_selected_edges();
-
-  size_t after = 0;
-  for (const auto& e : Sketch_access::get_edges(sketch)) if (e.node_idx_b.has_value()) ++after;
-  EXPECT_EQ(after, before + 1);
-
-  // Expect mirrored edge at x=-2
-  bool found = false;
-  for (const auto& e : Sketch_access::get_edges(sketch))
-  {
-    if (!e.node_idx_b.has_value()) continue;
-    const gp_Pnt2d& na = sketch.get_nodes()[e.node_idx_a];
-    const gp_Pnt2d& nb = sketch.get_nodes()[*e.node_idx_b];
-    if (std::abs(na.X() + 2.0) < Precision::Confusion() &&
-        std::abs(nb.X() + 2.0) < Precision::Confusion() &&
-        std::abs(na.Y() - 1.0) < Precision::Confusion() &&
-        std::abs(nb.Y() - 3.0) < Precision::Confusion())
-    {
-      found = true;
-      break;
-    }
-  }
-  EXPECT_TRUE(found) << "Mirrored edge should appear on the left of vertical axis";
-}
-
-// Test that mirroring an arc (circle arc pair) works
-TEST_F(Sketch_test, MirrorSelectedEdges_Arc)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_operation_axis);
-
-  // Horizontal axis
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 0.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(10.0, 0.0)));
-
-  // Add a simple arc above the axis (using three points)
-  // Arc from (-1,1) through (0,2) to (1,1) — a bump above x axis
-  gp_Pnt2d a(-1.0, 1.0);
-  gp_Pnt2d b(0.0, 2.0);
-  gp_Pnt2d c(1.0, 1.0);
-  Sketch_access::add_arc_circle_(sketch, a, b, c);
-
-  // For arcs, the shp is shared; select the shp of one of the arc edges
-  AIS_Shape_ptr arc_shp;
-  for (const auto& e : Sketch_access::get_edges(sketch))
-  {
-    if (sketch_edge_is_arc(e))
-    {
-      arc_shp = e.shp;
-      break;
-    }
-  }
-  ASSERT_FALSE(arc_shp.IsNull()) << "Arc should have been added with a shp";
-
-  auto& cctx = view().ctx();
-  cctx.ClearSelected(true);
-  cctx.AddOrRemoveSelected(arc_shp, true);
-
-  size_t before = 0;
-  for (const auto& e : Sketch_access::get_edges(sketch)) if (e.node_idx_b.has_value()) ++before;
-
-  sketch.mirror_selected_edges();
-
-  // Mirroring an arc pair should add another arc pair below
-  size_t after = 0;
-  for (const auto& e : Sketch_access::get_edges(sketch)) if (e.node_idx_b.has_value()) ++after;
-  // Expect +1 edge for the mirrored arc
-  EXPECT_EQ(after, before + 1) << "Mirroring an arc should add one new arc edge";
-
-  // Spot-check that a mirrored point exists below the axis (e.g. around y=-1 or so)
-  bool found_symmetric = false;
-  for (const auto& e : Sketch_access::get_edges(sketch))
-  {
-    if (!sketch_edge_is_arc(e))
-      continue;
-    const gp_Pnt2d& pa = sketch.get_nodes()[e.node_idx_a];
-    if (std::abs(pa.Y() + 1.0) < 0.5)  // rough, symmetric to +1
-    {
-      found_symmetric = true;
-      break;
-    }
-  }
-  EXPECT_TRUE(found_symmetric) << "A mirrored arc point should exist below the axis";
-}
-
-// Unit tests for the revolve tool (Sketch::revolve_selected).
-// Revolve requires a defined operation axis + selected edges or faces.
-// It returns a Shp_rslt (success or error). On success it produces a revolved 3D shape via BRepPrimAPI_MakeRevol.
-
-// Error case: no edges/faces selected (after axis is set)
-TEST_F(Sketch_test, RevolveSelected_NoSelection)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_operation_axis);
-
-  // Define a simple horizontal axis
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 0.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(5.0, 0.0)));
-
-  EXPECT_TRUE(sketch.has_operation_axis());
-
-  // Call without any selection
-  Shp_rslt res = sketch.revolve_selected(2 * std::numbers::pi);
-
-  EXPECT_FALSE(res.has_value());
-  EXPECT_EQ(res.message(), "No selected faces or edges.");
-}
-
-// Basic success case: revolve a straight edge profile around the axis (360 deg full revolution).
-// Uses edge selection (common for profile-based revolution).
-TEST_F(Sketch_test, RevolveSelected_SimpleEdgeProfile)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_operation_axis);
-
-  // Horizontal axis at y=0
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 0.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(10.0, 0.0)));
-
-  // Add a profile edge offset from the axis (a line segment parallel-ish, at x~2, y from 1 to 2)
-  // Revolving this around the x-axis will create a surface of revolution.
-  gp_Pnt2d p1(2.0, 1.0);
-  gp_Pnt2d p2(2.0, 2.0);
-  Sketch_access::add_edge_(sketch, p1, p2);
-
-  // Select the edge shp (same mechanism used by mirror tests and the real UI selection)
-  AIS_Shape_ptr edge_shp;
-  for (auto rit = Sketch_access::get_edges(sketch).rbegin(); rit != Sketch_access::get_edges(sketch).rend(); ++rit)
-  {
-    if (rit->node_idx_b.has_value())
-    {
-      edge_shp = rit->shp;
-      break;
-    }
-  }
-  ASSERT_FALSE(edge_shp.IsNull());
-
-  auto& cctx = view().ctx();
-  cctx.ClearSelected(true);
-  cctx.AddOrRemoveSelected(edge_shp, true);
-
-  Shp_rslt res = sketch.revolve_selected(2 * std::numbers::pi);
-
-  EXPECT_TRUE(res.has_value()) << "Revolve should succeed with a valid axis + selected edge. Message: " << res.message();
-  ASSERT_TRUE(res.has_value());
-
-  const TopoDS_Shape& actual = (*res)->Shape();
-  EXPECT_FALSE(actual.IsNull()) << "Revolved shape must not be null";
-
-  // === Stronger "oracle" test: rebuild the exact same revolve using low-level OCCT ===
-  // This verifies that Sketch::revolve_selected produces the *identical* geometry
-  // as a direct BRepPrimAPI_MakeRevol call with the extracted profile + axis.
-
-  // 1. Build the profile compound (same edges the sketch would have selected)
-  TopoDS_Compound profile;
-  BRep_Builder bld;
-  bld.MakeCompound(profile);
-  bld.Add(profile, edge_shp->Shape());   // the edge we selected above
-
-  // 2. Recreate the axis using the exact same two points we used to define the operation axis.
-  //    (Only direction + a point on the line matter; length is irrelevant.)
-  gp_Pnt axA(0.0, 0.0, 0.0);
-  gp_Pnt axB(10.0, 0.0, 0.0);
-  gp_Dir dir((axB.XYZ() - axA.XYZ()).Normalized());
-  gp_Ax1 revolAxis(axA, dir);
-
-  BRepPrimAPI_MakeRevol expectedMaker(profile, revolAxis, 2 * std::numbers::pi);
-  TopoDS_Shape expected = try_make_solid(expectedMaker.Shape());
-
-  EXPECT_FALSE(expected.IsNull());
-
-  // Stronger oracle comparison using mass properties + bounding box.
-  // This verifies that the geometry produced by the Sketch revolve wrapper
-  // is the same as a direct low-level BRepPrimAPI_MakeRevol call.
-  GProp_GProps gActual, gExpected;
-  BRepGProp::VolumeProperties(actual, gActual);
-  BRepGProp::VolumeProperties(expected, gExpected);
-
-  EXPECT_NEAR(gActual.Mass(), gExpected.Mass(), 1e-8)
-      << "Revolved volume should match direct MakeRevol";
-
-  Bnd_Box boxA, boxE;
-  BRepBndLib::Add(actual, boxA);
-  BRepBndLib::Add(expected, boxE);
-
-  double axmin,aymin,azmin,axmax,aymax,azmax;
-  double exmin,eymin,ezmin,exmax,eymax,ezmax;
-  boxA.Get(axmin,aymin,azmin,axmax,aymax,azmax);
-  boxE.Get(exmin,eymin,ezmin,exmax,eymax,ezmax);
-
-  EXPECT_NEAR(axmin, exmin, 1e-8);
-  EXPECT_NEAR(aymin, eymin, 1e-8);
-  EXPECT_NEAR(azmin, ezmin, 1e-8);
-  EXPECT_NEAR(axmax, exmax, 1e-8);
-  EXPECT_NEAR(aymax, eymax, 1e-8);
-  EXPECT_NEAR(azmax, ezmax, 1e-8);
-
-  // It will typically be a Face (surface of revolution) for an open edge profile.
-  // We don't assert Solid here because that usually requires a closed face profile.
-}
-
-// Revolve a closed edge profile (rectangle) by selecting its boundary edges.
-// This exercises the selected_edges path in revolve_selected (multiple edges in compound).
-// Revolving a closed profile 360° around an external axis produces a solid of revolution.
-TEST_F(Sketch_test, RevolveSelected_ClosedEdgeProfile)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_operation_axis);
-
-  // Axis
-  sketch.add_sketch_pt(ScreenCoords(dvec2(0.0, 0.0)));
-  sketch.add_sketch_pt(ScreenCoords(dvec2(10.0, 0.0)));
-
-  // Simple closed rectangular profile offset from the axis (will form a face too, but we select edges).
-  gp_Pnt2d p0(1.0, 1.0);
-  gp_Pnt2d p1(3.0, 1.0);
-  gp_Pnt2d p2(3.0, 2.0);
-  gp_Pnt2d p3(1.0, 2.0);
-
-  Sketch_access::add_edge_(sketch, p0, p1);
-  Sketch_access::add_edge_(sketch, p1, p2);
-  Sketch_access::add_edge_(sketch, p2, p3);
-  Sketch_access::add_edge_(sketch, p3, p0);
-
-  // Select all four profile edges (the revolve code will compound their shapes)
-  auto& cctx = view().ctx();
-  cctx.ClearSelected(true);
-
-  size_t selected_count = 0;
-  for (const auto& e : Sketch_access::get_edges(sketch))
-  {
-    if (e.node_idx_b.has_value())
-    {
-      // Only select the ones we just added for the profile (simple heuristic by x/y range)
-      const gp_Pnt2d& na = sketch.get_nodes()[e.node_idx_a];
-      const gp_Pnt2d& nb = sketch.get_nodes()[*e.node_idx_b];
-      if (na.X() >= 0.5 && na.X() <= 3.5 && nb.X() >= 0.5 && nb.X() <= 3.5)
-      {
-        cctx.AddOrRemoveSelected(e.shp, true);
-        ++selected_count;
-      }
-    }
-  }
-  EXPECT_GE(selected_count, 2) << "Should have selected at least a couple of profile edges";
-
-  Shp_rslt res = sketch.revolve_selected(2 * std::numbers::pi);
-
-  EXPECT_TRUE(res.has_value()) << "Revolve of closed edge profile should succeed. Message: " << res.message();
-  ASSERT_TRUE(res.has_value());
-
-  const TopoDS_Shape& actual = (*res)->Shape();
-  EXPECT_FALSE(actual.IsNull()) << "Revolved shape must not be null";
-
-  // === Oracle comparison: rebuild the identical revolve using the same inputs ===
-  TopoDS_Compound profile;
-  BRep_Builder bld;
-  bld.MakeCompound(profile);
-
-  // Collect exactly the profile edges we added for this test
-  for (const auto& e : Sketch_access::get_edges(sketch))
-  {
-    if (e.node_idx_b.has_value())
-    {
-      const gp_Pnt2d& na = sketch.get_nodes()[e.node_idx_a];
-      const gp_Pnt2d& nb = sketch.get_nodes()[*e.node_idx_b];
-      if (na.X() >= 0.5 && na.X() <= 3.5 && nb.X() >= 0.5 && nb.X() <= 3.5)
-        bld.Add(profile, e.shp->Shape());
-    }
-  }
-
-  // Reconstruct the axis from the same points we used when creating the operation axis
-  gp_Pnt axA(0.0, 0.0, 0.0);
-  gp_Pnt axB(10.0, 0.0, 0.0);
-  gp_Dir dir((axB.XYZ() - axA.XYZ()).Normalized());
-  gp_Ax1 revolAxis(axA, dir);
-
-  BRepPrimAPI_MakeRevol expectedMaker(profile, revolAxis, 2 * std::numbers::pi);
-  TopoDS_Shape expected = try_make_solid(expectedMaker.Shape());
-
-  EXPECT_FALSE(expected.IsNull());
-
-  GProp_GProps gActual, gExpected;
-  BRepGProp::VolumeProperties(actual, gActual);
-  BRepGProp::VolumeProperties(expected, gExpected);
-
-  EXPECT_NEAR(gActual.Mass(), gExpected.Mass(), 1e-8)
-      << "Revolved volume should match direct MakeRevol";
-
-  Bnd_Box boxA, boxE;
-  BRepBndLib::Add(actual, boxA);
-  BRepBndLib::Add(expected, boxE);
-
-  double axmin,aymin,azmin,axmax,aymax,azmax;
-  double exmin,eymin,ezmin,exmax,eymax,ezmax;
-  boxA.Get(axmin,aymin,azmin,axmax,aymax,azmax);
-  boxE.Get(exmin,eymin,ezmin,exmax,eymax,ezmax);
-
-  EXPECT_NEAR(axmin, exmin, 1e-8);
-  EXPECT_NEAR(aymin, eymin, 1e-8);
-  EXPECT_NEAR(azmin, ezmin, 1e-8);
-  EXPECT_NEAR(axmax, exmax, 1e-8);
-  EXPECT_NEAR(aymax, eymax, 1e-8);
-  EXPECT_NEAR(azmax, ezmax, 1e-8);
-
-  // For a closed profile revolved 360°, we expect a solid of revolution.
-  EXPECT_EQ(actual.ShapeType(), TopAbs_SOLID) << "Closed edge profile revolved 360 deg should be a solid";
-}
-
-// Test that split edges have midpoints for snapping
-// This test verifies the fix for the bug where edges split by their midpoint
-// don't have midpoints to snap to
-// don't have midpoints to snap to
-// This test verifies the fix for the bug where edges split by their midpoint
-// don't have midpoints to snap to
-TEST_F(Sketch_test, SplitEdge_HasMidpoints)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  // Enable midpoints for this test (it verifies midpoint creation and split-mid behavior)
-  sketch.set_add_mid_pt_edges(true);
-
-  // Create a simple horizontal edge
-  gp_Pnt2d pt_left(0.0, 0.0);
-  gp_Pnt2d pt_right(20.0, 0.0);
-  
-  // Add the edge - this will automatically create a midpoint node at (10, 0)
-  Sketch_access::add_edge_(sketch, pt_left, pt_right);
-
-  // Count initial nodes and edges
-  size_t initial_node_count = sketch.get_nodes().size();
-  size_t initial_edge_count = 0;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-    if (edge.node_idx_b.has_value())
-      initial_edge_count++;
-
-  EXPECT_EQ(initial_edge_count, 1) << "Should have 1 edge initially";
-  EXPECT_EQ(initial_node_count, 3) << "Should have 3 nodes initially (left, right, midpoint)";
-
-  // Find the midpoint of the initial edge
-  std::optional<size_t> initial_midpoint_idx;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-  {
-    if (edge.node_idx_b.has_value() && edge.node_idx_mid.has_value())
-    {
-      initial_midpoint_idx = edge.node_idx_mid;
-      break;
-    }
-  }
-  ASSERT_TRUE(initial_midpoint_idx.has_value()) << "Initial edge should have a midpoint";
-
-  // Verify the midpoint is at (10, 0)
-  const gp_Pnt2d& midpoint = sketch.get_nodes()[*initial_midpoint_idx];
-  EXPECT_NEAR(midpoint.X(), 10.0, Precision::Confusion());
-  EXPECT_NEAR(midpoint.Y(), 0.0, Precision::Confusion());
-
-  // Set mode to add edges (set_mode syncs mids from GUI prefs; re-enable for this test)
-  gui().set_mode(Mode::Sketch_add_edge);
-  Sketch::set_add_mid_pt_edges(true);
-
-  // Create a new edge that snaps to the midpoint, which will split the original edge
-  // Start from the midpoint
-  ScreenCoords screen_coords_mid(dvec2(10.0, 0.0));
-  sketch.add_sketch_pt(screen_coords_mid);
-
-  // Add another point above the midpoint to create a vertical edge
-  gp_Pnt2d pt_above(10.0, 10.0);
-  ScreenCoords screen_coords_above(dvec2(pt_above.X(), pt_above.Y()));
-  sketch.add_sketch_pt(screen_coords_above);
-
-  // Finalize the edge - this should trigger the edge splitting
-  sketch.finalize_elm();
-
-  // Count edges after splitting
-  size_t final_edge_count = 0;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-    if (edge.node_idx_b.has_value())
-      final_edge_count++;
-
-  // Should now have 3 edges: two from the split horizontal edge, plus the new vertical edge
-  EXPECT_EQ(final_edge_count, 3) << "Should have 3 edges after splitting (2 horizontal + 1 vertical)";
-
-  // Verify that both split edges have midpoints
-  std::vector<std::optional<size_t>> split_edge_midpoints;
-  for (const auto& edge : Sketch_access::get_edges(sketch))
-  {
-    if (!edge.node_idx_b.has_value())
-      continue;
-
-    const gp_Pnt2d& pt_a = sketch.get_nodes()[edge.node_idx_a];
-    const gp_Pnt2d& pt_b = sketch.get_nodes()[*edge.node_idx_b];
-
-    // Check if this is one of the horizontal split edges (y = 0)
-    if (std::abs(pt_a.Y()) < Precision::Confusion() && 
-        std::abs(pt_b.Y()) < Precision::Confusion())
-    {
-      // This is a horizontal edge, verify it has a midpoint
-      EXPECT_TRUE(edge.node_idx_mid.has_value()) 
-          << "Split horizontal edge should have a midpoint";
-      
-      if (edge.node_idx_mid.has_value())
-      {
-        split_edge_midpoints.push_back(edge.node_idx_mid);
-        
-        // Verify the midpoint is correctly positioned
-        const gp_Pnt2d& edge_midpoint = sketch.get_nodes()[*edge.node_idx_mid];
-        double expected_x = (pt_a.X() + pt_b.X()) / 2.0;
-        EXPECT_NEAR(edge_midpoint.X(), expected_x, Precision::Confusion())
-            << "Midpoint should be at the center of the edge";
-        EXPECT_NEAR(edge_midpoint.Y(), 0.0, Precision::Confusion())
-            << "Midpoint should have y=0";
-        
-        // Verify the midpoint is marked as a midpoint
-        EXPECT_TRUE(sketch.get_nodes()[*edge.node_idx_mid].midpoint)
-            << "Midpoint node should be marked as midpoint";
-      }
-    }
-  }
-
-  // Should have found two split edges with midpoints
-  EXPECT_EQ(split_edge_midpoints.size(), 2) 
-      << "Should have found 2 horizontal split edges with midpoints";
-
-  // Verify the split edge midpoints are at (5, 0) and (15, 0)
-  std::vector<double> midpoint_x_coords;
-  for (const auto& mid_idx : split_edge_midpoints)
-  {
-    if (mid_idx.has_value())
-    {
-      const gp_Pnt2d& mp = sketch.get_nodes()[*mid_idx];
-      midpoint_x_coords.push_back(mp.X());
-    }
-  }
-  
-  std::sort(midpoint_x_coords.begin(), midpoint_x_coords.end());
-  ASSERT_EQ(midpoint_x_coords.size(), 2);
-  EXPECT_NEAR(midpoint_x_coords[0], 5.0, Precision::Confusion()) 
-      << "First split edge should have midpoint at x=5";
-  EXPECT_NEAR(midpoint_x_coords[1], 15.0, Precision::Confusion()) 
-      << "Second split edge should have midpoint at x=15";
-}
-
 TEST_F(Sketch_test, AddNode_splits_linear_edge_interior)
 {
   gp_Pln default_plane(gp::Origin(), gp::DZ());
@@ -3416,19 +794,7 @@ TEST_F(Sketch_test, AddNode_splits_linear_edge_interior)
 
 TEST_F(Sketch_test, AddNode_off_edge_adds_node_only)
 {
-  struct Headless_guard
-  {
-    Occt_view& m_v;
-    explicit Headless_guard(Occt_view& v)
-        : m_v(v)
-    {
-      View_access::set_headless(m_v, true);
-    }
-    ~Headless_guard()
-    {
-      View_access::set_headless(m_v, false);
-    }
-  } guard(view());
+  Headless_guard guard(view());
 
   gp_Pln default_plane(gp::Origin(), gp::DZ());
   Sketch sketch("TestSketch", view(), default_plane);
@@ -3452,21 +818,11 @@ TEST_F(Sketch_test, AddNode_off_edge_adds_node_only)
 }
 
 // Pick slightly off a segment in plane space; should snap onto the edge and split so the node stays snappable.
+
+// Pick slightly off a segment in plane space; should snap onto the edge and split so the node stays snappable.
 TEST_F(Sketch_test, AddNode_near_edge_snaps_onto_segment_and_splits)
 {
-  struct Headless_guard
-  {
-    Occt_view& m_v;
-    explicit Headless_guard(Occt_view& v)
-        : m_v(v)
-    {
-      View_access::set_headless(m_v, true);
-    }
-    ~Headless_guard()
-    {
-      View_access::set_headless(m_v, false);
-    }
-  } guard(view());
+  Headless_guard guard(view());
 
   gp_Pln default_plane(gp::Origin(), gp::DZ());
   Sketch sketch("TestSketch", view(), default_plane);
@@ -3530,19 +886,7 @@ TEST_F(Sketch_test, AddNode_distance_enter_undo)
 
 TEST_F(Sketch_test, AddNode_anchor_then_click_undo)
 {
-  struct Headless_guard
-  {
-    Occt_view& m_v;
-    explicit Headless_guard(Occt_view& v)
-        : m_v(v)
-    {
-      View_access::set_headless(m_v, true);
-    }
-    ~Headless_guard()
-    {
-      View_access::set_headless(m_v, false);
-    }
-  } guard(view());
+  Headless_guard guard(view());
 
   gp_Pln default_plane(gp::Origin(), gp::DZ());
   Sketch sketch("TestSketch", view(), default_plane);
@@ -3569,35 +913,9 @@ TEST_F(Sketch_test, AddNode_anchor_then_click_undo)
   EXPECT_EQ(count_permanent_nodes(sketch), 2u) << "Redo should restore the click-placed node";
 }
 
-TEST_F(Sketch_test, NewFile_clears_undo_stacks)
-{
-  gp_Pln default_plane(gp::Origin(), gp::DZ());
-  Sketch sketch("TestSketch", view(), default_plane);
-
-  gui().set_mode(Mode::Sketch_add_node);
-  sketch.add_sketch_pt(ScreenCoords(dvec2(10.0, 10.0)));
-  ASSERT_GT(view().undo_stack_size(), 0u);
-
-  view().new_file();
-  EXPECT_EQ(view().undo_stack_size(), 0u);
-  EXPECT_EQ(view().redo_stack_size(), 0u);
-}
-
 TEST_F(Sketch_test, DimensionTool_picks_sketch_origin)
 {
-  struct Headless_guard
-  {
-    Occt_view& m_v;
-    explicit Headless_guard(Occt_view& v)
-        : m_v(v)
-    {
-      View_access::set_headless(m_v, true);
-    }
-    ~Headless_guard()
-    {
-      View_access::set_headless(m_v, false);
-    }
-  } guard(view());
+  Headless_guard guard(view());
 
   gp_Pln default_plane(gp::Origin(), gp::DZ());
   Sketch sketch("TestSketch", view(), default_plane);
@@ -3607,132 +925,3 @@ TEST_F(Sketch_test, DimensionTool_picks_sketch_origin)
   EXPECT_TRUE(sketch.get_nodes()[*origin_pick].origin);
 }
 
-namespace
-{
-
-nlohmann::json minimal_sketch_json_with_underlay_b64(Ezy_asset_store& store)
-{
-  std::vector<uint8_t> rgba(16);
-  for (int i = 0; i < 4; ++i)
-  {
-    rgba[static_cast<std::size_t>(i) * 4u + 0] = 255;
-    rgba[static_cast<std::size_t>(i) * 4u + 1] = 0;
-    rgba[static_cast<std::size_t>(i) * 4u + 2] = 0;
-    rgba[static_cast<std::size_t>(i) * 4u + 3] = 255;
-  }
-
-  nlohmann::json j;
-  j["isCurrent"] = true;
-  j["name"]      = "UnderlaySketch";
-  j["plane"]     = nlohmann::json::object({{"origin", {{"x", 0.0}, {"y", 0.0}, {"z", 0.0}}},
-                                            {"normal", {{"x", 0.0}, {"y", 0.0}, {"z", 1.0}}},
-                                            {"xAxis", {{"x", 1.0}, {"y", 0.0}, {"z", 0.0}}}});
-  j["edges"]     = nlohmann::json::array();
-  j["nodes"]     = nlohmann::json::array();
-  j["arc_edges"] = nlohmann::json::array();
-  j["underlay"]  = nlohmann::json::object({{"rgba_b64", ezy_base64_encode(rgba)},
-                                             {"w", 2},
-                                             {"h", 2},
-                                             {"opacity", 0.88},
-                                             {"visible", true}});
-  (void)store;
-  return j;
-}
-
-} // namespace
-
-TEST_F(Sketch_test, EzyZipUnderlayRoundTrip)
-{
-  view().asset_store().clear();
-  nlohmann::json doc     = nlohmann::json::parse(view().to_json());
-  doc["sketches"]        = nlohmann::json::array({minimal_sketch_json_with_underlay_b64(view().asset_store())});
-  const std::string load = doc.dump();
-  view().load(load, false);
-
-  const auto sk = view().curr_sketch_shared();
-  ASSERT_TRUE(sk);
-  ASSERT_TRUE(sk->underlay().has_image());
-  EXPECT_EQ(sk->underlay().image_w(), 2);
-  EXPECT_EQ(sk->underlay().image_h(), 2);
-
-  const std::string manifest             = view().to_json();
-  const std::vector<uint8_t> zip_bytes   = pack_ezy(manifest, view().asset_store());
-  ASSERT_FALSE(zip_bytes.empty());
-  ASSERT_TRUE(is_ezy_zip(std::string(reinterpret_cast<const char*>(zip_bytes.data()), zip_bytes.size())));
-
-  auto unpacked = unpack_ezy(std::string(reinterpret_cast<const char*>(zip_bytes.data()), zip_bytes.size()));
-  ASSERT_TRUE(unpacked);
-  EXPECT_FALSE(unpacked->manifest_json.empty());
-  EXPECT_EQ(unpacked->assets.size(), 1u);
-
-  view().asset_store().clear();
-  for (auto& [id, data] : unpacked->assets)
-    view().asset_store().import_asset(id, std::move(data));
-
-  view().load(unpacked->manifest_json, false);
-  const auto loaded = view().curr_sketch_shared();
-  ASSERT_TRUE(loaded);
-  EXPECT_TRUE(loaded->underlay().has_image());
-  EXPECT_EQ(loaded->underlay().image_w(), 2);
-  EXPECT_EQ(loaded->underlay().image_h(), 2);
-}
-
-TEST_F(Sketch_test, EzyZipUnderlayDedup)
-{
-  view().asset_store().clear();
-  nlohmann::json sk1 = minimal_sketch_json_with_underlay_b64(view().asset_store());
-  nlohmann::json sk2 = minimal_sketch_json_with_underlay_b64(view().asset_store());
-  sk1["name"]        = "A";
-  sk2["name"]        = "B";
-  sk2["isCurrent"]   = false;
-  nlohmann::json doc;
-  doc["ezyFormat"]  = 3;
-  doc["shapes"]     = nlohmann::json::array();
-  doc["sketches"]   = nlohmann::json::array({sk1, sk2});
-  view().load(doc.dump(), false);
-
-  const std::string manifest         = view().to_json();
-  const std::vector<uint8_t> zip_bytes = pack_ezy(manifest, view().asset_store());
-  const auto unpacked = unpack_ezy(std::string(reinterpret_cast<const char*>(zip_bytes.data()), zip_bytes.size()));
-  ASSERT_TRUE(unpacked);
-  EXPECT_EQ(unpacked->assets.size(), 1u);
-
-  const nlohmann::json out = nlohmann::json::parse(unpacked->manifest_json);
-  std::set<std::string> asset_ids;
-  for (const nlohmann::json& sk : out["sketches"])
-  {
-    if (sk.contains("underlay") && sk["underlay"].is_object() && sk["underlay"].contains("asset"))
-      asset_ids.insert(sk["underlay"]["asset"].get<std::string>());
-  }
-  EXPECT_EQ(asset_ids.size(), 1u);
-  for (const nlohmann::json& sk : out["sketches"])
-  {
-    if (sk.contains("underlay") && sk["underlay"].is_object())
-      EXPECT_FALSE(sk["underlay"].contains("rgba_b64"));
-  }
-}
-
-TEST_F(Sketch_test, LegacyUnderlayRgbaB64Load)
-{
-  view().asset_store().clear();
-  const auto sk = Sketch_json::from_json(view(), minimal_sketch_json_with_underlay_b64(view().asset_store()));
-  ASSERT_TRUE(sk);
-  EXPECT_TRUE(sk->underlay().has_image());
-  const nlohmann::json saved = Sketch_json::to_json(*sk, view().asset_store());
-  EXPECT_TRUE(saved.contains("underlay"));
-  EXPECT_TRUE(saved["underlay"].contains("asset"));
-  EXPECT_FALSE(saved["underlay"].contains("rgba_b64"));
-}
-
-TEST_F(Sketch_test, UnderlayUndoSnapshotUsesAssetRef)
-{
-  view().asset_store().clear();
-  nlohmann::json doc = nlohmann::json::parse(view().to_json());
-  doc["sketches"]    = nlohmann::json::array({minimal_sketch_json_with_underlay_b64(view().asset_store())});
-  view().load(doc.dump(), false);
-
-  view().push_undo_snapshot();
-  const std::string snap = view().to_json();
-  EXPECT_EQ(snap.find("rgba_b64"), std::string::npos);
-  EXPECT_NE(snap.find("\"asset\""), std::string::npos);
-}
